% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{AddMetaData-v4}
\alias{AddMetaData-v4}
\alias{AddMetaData.Assay}
\title{Add in metadata associated with either cells or features.}
\usage{
\method{AddMetaData}{Assay}(object, metadata, col.name = NULL)
}
\arguments{
\item{object}{An object}

\item{metadata}{A vector, list, or data.frame with metadata to add}

\item{col.name}{A name for meta data if not a named list or data.frame}
}
\value{
\code{object} with metadata added
}
\description{
Adds additional data to the object. Can be any piece of information
associated with a cell (examples include read depth, alignment rate,
experimental batch, or subpopulation identity) or feature (ENSG name,
variance). To add cell level information, add to the Seurat object. If adding
feature-level metadata, add to the Assay object (e.g. \code{object[["RNA"]]})
}
\note{
This is the documentation for \code{AddMetaData} for
v4; for the latest documentation, please see
\code{\link{AddMetaData}}
}
\examples{
cluster_letters <- LETTERS[Idents(object = pbmc_small)]
names(cluster_letters) <- colnames(x = pbmc_small)
pbmc_small <- AddMetaData(
  object = pbmc_small,
  metadata = cluster_letters,
  col.name = 'letter.idents'
)
head(x = pbmc_small[[]])

}
\seealso{
\code{\link{AddMetaData}}
}
