% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Angles}
\alias{Angles}
\alias{Degrees}
\alias{Radians}
\title{Radian/Degree Conversions}
\usage{
Degrees(rad)

Radians(deg)
}
\arguments{
\item{rad}{Angle in radians}

\item{deg}{Angle in degrees}
}
\value{
\code{Degrees}: \code{rad} in degrees

\code{Radians}: \code{deg} in radians
}
\description{
Convert degrees to radians and vice versa
}
\examples{
Degrees(pi)

Radians(180)

}
\seealso{
Other angles: 
\code{\link{PolyVtx}()}
}
\concept{angles}
\concept{utils}
\keyword{internal}
