% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.Centroids}
\alias{as.Centroids}
\alias{as.Segmentation}
\title{Convert Segmentation Layers}
\usage{
as.Centroids(x, nsides = NULL, radius = NULL, theta = NULL, ...)

as.Segmentation(x, ...)
}
\arguments{
\item{x}{An object}

\item{nsides}{The number of sides to represent cells/spots; pass
\code{\link[base]{Inf}} to plot as circles}

\item{radius}{Radius of shapes when plotting}

\item{theta}{Angle to adjust shapes when plotting}
}
\value{
\code{as.Centroids}: A \code{\link{Centroids}} object

\code{as.Segmentation}: A \code{\link{Segmentation}} object
}
\description{
Convert Segmentation Layers
}
