% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{as.list.SeuratCommand}
\alias{as.list.SeuratCommand}
\title{Coerce a \code{SeuratCommand} to a list}
\usage{
\method{as.list}{SeuratCommand}(x, complete = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{SeuratCommand}} object}

\item{complete}{Include slots besides just parameters
(eg. call string, name, timestamp)}
}
\value{
A list with the parameters and, if \code{complete = TRUE},
the call string, name, and timestamp
}
\description{
Coerce a \code{SeuratCommand} to a list
}
\examples{
cmd <- pbmc_small[["NormalizeData.RNA"]]
as.list(cmd)
as.list(cmd, complete = TRUE)

}
\seealso{
Command log object and interaction methods
\code{\link{$.SeuratCommand}()},
\code{\link{.DollarNames.SeuratCommand}()},
\code{\link{LogSeuratCommand}()},
\code{\link{SeuratCommand-class}},
\code{\link{[.SeuratCommand}()}
}
\concept{command}
