% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/neighbor.R
\name{as.Neighbor}
\alias{as.Neighbor}
\alias{as.Neighbor.Graph}
\title{Coerce to a \code{Neighbor} Object}
\usage{
as.Neighbor(x, ...)

\method{as.Neighbor}{Graph}(x, ...)
}
\arguments{
\item{x}{An object to convert to \code{\link{Neighbor}}}

\item{...}{Arguments passed to other methods}
}
\value{
A \code{\link{Neighbor}} object
}
\description{
Convert objects to \code{\link{Neighbor}} objects
}
\concept{neighbor}
