% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.sparse}
\alias{as.sparse}
\title{Convert between data frames and sparse matrices}
\usage{
as.sparse(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{as.sparse}: A sparse representation of the input data
}
\description{
Convert between data frames and sparse matrices
}
