% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{Assay-methods}
\alias{Assay-methods}
\alias{[.Assay}
\alias{[[.Assay}
\alias{dim.Assay}
\alias{dimnames.Assay}
\alias{merge.Assay}
\alias{subset.Assay}
\title{\code{Assay} Methods}
\usage{
\method{[}{Assay}(x, i, j, ...)

\method{[[}{Assay}(x, i, ..., drop = FALSE)

\method{dim}{Assay}(x)

\method{dimnames}{Assay}(x)

\method{merge}{Assay}(x = NULL, y = NULL, add.cell.ids = NULL, merge.data = TRUE, ...)

\method{subset}{Assay}(x, cells = NULL, features = NULL, ...)
}
\arguments{
\item{i, features}{For \code{[[}: metadata names; for all other methods,
feature names or indices}

\item{j, cells}{Cell names or indices}

\item{...}{Arguments passed to other methods}

\item{drop}{See \code{\link[base]{drop}}}

\item{y}{A vector or list of one or more objects to merge}

\item{add.cell.ids}{A character vector of \code{length(x = c(x, y))}; appends
the corresponding values to the start of each objects' cell names}

\item{merge.data}{Merge the data slots instead of just merging the counts
(which requires renormalization); this is recommended if the same normalization
approach was applied to all objects}

\item{object, x}{An \code{\link{Assay}} object}
}
\value{
\code{[}: The \code{data} slot for features \code{i} and cells
\code{j}

\code{[[}: The feature-level metadata for \code{i}

\code{dim}: The number of features (\code{nrow}) and cells
\code{ncol}

\code{dimnames}: Feature (row) and cell (column) names

\code{merge}: Merged object

\code{subset}: A subsetted object
}
\description{
Methods for \code{\link{Assay}} for generics defined in other packages
}
\section{Functions}{
\itemize{
\item \code{[.Assay}: Data accessor

\item \code{[[.Assay}: Metadata accessor

\item \code{dim.Assay}: Number of cells and features

\item \code{dimnames.Assay}: Cell- and feature-names

\item \code{merge.Assay}: Merge objects

\item \code{subset.Assay}: Subset object
}}

