% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\docType{class}
\name{Assay5-class}
\alias{Assay5-class}
\alias{Assay5}
\title{The v5 \code{Assay} Object}
\description{
The v5 \code{Assay} is the typical \code{Assay} class used in \pkg{Seurat}
v5; ...
}
\seealso{
\code{\link{StdAssay-class}}
}
