% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{Assay5-validity}
\alias{Assay5-validity}
\title{V5 Assay Validity}
\description{
Validation of \code{Assay5} objects is handled by
\code{\link[methods]{validObject}}
}
\section{Layer Validation}{

blah
}

\section{Key Validation}{

Keys must be a one-length character vector; a key must be composed of one
of the following:
\itemize{
 \item An empty string (eg. \dQuote{\code{''}}) where \code{nzchar() == 0}
 \item An string composed of one or more alphanumeric values
 (both lower- and upper-case) that ends with an underscore
 (\dQuote{\code{_}}); the first character must be a letter
}
Keys that are not empty strings are validated with the regex
\dQuote{\code{\Sexpr[stage=build]{SeuratObject:::.KeyPattern()}}}
}

\seealso{
\code{\link[methods]{validObject}}

v5 Assay object, validity, and interaction methods:
\code{\link{Assay5-class}},
\code{\link{[.Assay5}()},
\code{\link{[[.Assay5}()},
\code{\link{dim.Assay5}()},
\code{\link{dimnames.Assay5}()},
\code{\link{merge.Assay5}()},
\code{\link{split.Assay5}()},
\code{\link{subset.Assay5}()}
}
\concept{assay5}
