% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{AssayData-StdAssay}
\alias{AssayData-StdAssay}
\alias{GetAssayData.StdAssay}
\alias{SetAssayData.StdAssay}
\title{Get and Set Assay Data}
\usage{
\method{GetAssayData}{StdAssay}(object, layer = NULL, slot = deprecated(), ...)

\method{SetAssayData}{StdAssay}(object, layer, slot = deprecated(), new.data, ...)
}
\arguments{
\item{object}{An object}

\item{layer}{Name of layer to get or set}

\item{slot}{\Sexpr[stage=build,results=rd]{lifecycle::badge("deprecated")} Specific assay data to get or set}

\item{...}{Arguments passed to other methods}

\item{new.data}{New assay data to add}
}
\value{
\code{GetAssayData}: returns the specified assay data

\code{SetAssayData}: \code{object} with the assay data set
}
\description{
General accessor and setter functions for \code{\link{Assay}} objects.
\code{GetAssayData} can be used to pull information from any of the
expression matrices (eg. \dQuote{counts}, \dQuote{data}, or
\dQuote{scale.data}). \code{SetAssayData} can be used to replace one of these
expression matrices
}
\section{Lifecycle}{


\Sexpr[stage=build,results=rd]{lifecycle::badge("superseded")}



\code{GetAssayData} and \code{SetAssayData} have been superseded. To fetch
expression matrices, use \code{\link{LayerData}}; to set expression data,
use \code{\link{LayerData<-}}
}

\keyword{internal}
