% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{CalcN}
\alias{CalcN}
\title{Calculate nCount and nFeature}
\usage{
CalcN(object)
}
\arguments{
\item{object}{An \code{\link{Assay}} object}
}
\value{
A named list with nCount and nFeature
}
\description{
Calculate nCount and nFeature
}
\examples{
\donttest{
calcn <- SeuratObject:::CalcN(pbmc_small[["RNA"]])
head(as.data.frame(calcn))
}

}
\keyword{internal}
