% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{$.SeuratCommand}
\alias{$.SeuratCommand}
\title{Command Log Parameter Access}
\usage{
\method{$}{SeuratCommand}(x, i, ...)
}
\arguments{
\item{x}{A \code{\link{SeuratCommand}} object}

\item{i}{A parameter name}
}
\value{
The value for parameter \code{i}
}
\description{
Pull parameter values from a \code{\link{SeuratCommand}} object
}
\examples{
cmd <- pbmc_small[["NormalizeData.RNA"]]
cmd$normalization.method

}
\seealso{
Command log object and interaction methods
\code{\link{.DollarNames.SeuratCommand}()},
\code{\link{LogSeuratCommand}()},
\code{\link{SeuratCommand-class}},
\code{\link{[.SeuratCommand}()},
\code{\link{as.list.SeuratCommand}()}
}
\concept{command}
