% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{$.StdAssay}
\alias{$.StdAssay}
\alias{$<-.StdAssay}
\title{Layer Data}
\usage{
\method{$}{StdAssay}(x, i)

\method{$}{StdAssay}(x, i) <- value
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{i}{Name of layer data to get or set}

\item{value}{A matrix-like object to add as a new layer}
}
\value{
{$}: Layer data for layer \code{i}

\code{$<-}: \code{x} with layer data \code{value} saved as \code{i}
}
\description{
Get and set layer data
}
\seealso{
v5 Standard Assay object, validity, and interaction methods
\code{\link{.DollarNames.StdAssay}()},
\code{\link{StdAssay-class}},
\code{\link{StdAssay-validity}},
\code{\link{[.StdAssay}()},
\code{\link{[[.StdAssay}()},
\code{\link{dim.StdAssay}()},
\code{\link{dimnames.StdAssay}()},
\code{\link{show,StdAssay-method}},
\code{\link{split.StdAssay}()},
\code{\link{subset.StdAssay}()}
}
\concept{stdassay}
\keyword{internal}
