% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay5.R
\name{CastAssay}
\alias{CastAssay}
\alias{CastAssay.Assay5}
\title{Cast Assay Layers}
\usage{
CastAssay(object, to, ...)

\method{CastAssay}{Assay5}(object, to, layers = NA, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object}

\item{to}{Either a class name or a function that takes a layer and returns
the same layer as a new class}

\item{...}{If \code{to} is a function, arguments passed to \code{to}}

\item{layers}{A vector of layers to cast; defaults to all layers}

\item{verbose}{Show progress updates}
}
\value{
\code{object} with the layers cast to class specified by \code{to}
}
\description{
Cast layers in v5 assays to other classes
}
