% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{Cells-StdAssay}
\alias{Cells-StdAssay}
\alias{Cells.StdAssay}
\alias{Features.StdAssay}
\title{Cell and Feature Names}
\usage{
\method{Cells}{StdAssay}(x, layer = NULL, simplify = TRUE, ...)

\method{Features}{StdAssay}(x, layer = NULL, simplify = TRUE, ...)
}
\arguments{
\item{x}{An object}

\item{layer}{Layer to pull cells/features for; defaults to default layer;
if \code{NA}, returns all cells for the assay}

\item{simplify}{Simplify the cell/feature names into a single vector; if
\code{FALSE}, separates each cell/feature names by layer}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Cell}: A vector of cell names

\code{Features}: A vector of feature names
}
\description{
Get the cell and feature names of an object
}
\keyword{internal}
