% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/default.R, R/dimreduc.R,
%   R/neighbor.R
\name{Cells}
\alias{Cells}
\alias{Cells.default}
\alias{Cells.DimReduc}
\alias{Cells.Neighbor}
\title{Get cells present in an object}
\usage{
Cells(x)

\method{Cells}{default}(x)

\method{Cells}{DimReduc}(x)

\method{Cells}{Neighbor}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A vector of cell names
}
\description{
Get cells present in an object
}
\examples{
Cells(x = pbmc_small)

}
\concept{data-access}
