% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids.R
\docType{class}
\name{Centroids-class}
\alias{Centroids-class}
\title{The \code{Centroids} Class}
\description{
The \code{Centroids} Class
}
\section{Slots}{

\describe{
\item{\code{cells}}{(\code{\link[base:character]{character [n]}}) A vector of cell
names; there should be as many cell names as there are points and no
duplicate names}

\item{\code{nsides}}{(\code{\link[base:integer]{integer [1L]}}) The number of sides
to draw when plotting centroids; must be either \code{0L} for circles or
greater than 3}

\item{\code{radius}}{(\code{\link[base:numeric]{numeric [1L]}}) The radius of the
shape when plotting the centroids}

\item{\code{theta}}{(\code{\link[base:numeric]{numeric [1L]}}) The angle in degrees
to adjust the shape when plotting the centroids}
}}

\seealso{
\code{Centroids} methods: \code{\link{Centroids-methods}}

Segmentation layer classes:
\code{\link{Centroids-methods}},
\code{\link{Molecules-class}},
\code{\link{Molecules-methods}},
\code{\link{Segmentation-class}},
\code{\link{Segmentation-methods}}
}
\concept{segmentation}
