% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckDots}
\alias{CheckDots}
\title{Check the use of dots}
\usage{
CheckDots(..., fxns = NULL)
}
\arguments{
\item{...}{Arguments passed to a function that fall under \code{...}}

\item{fxns}{A list/vector of functions or function names}
}
\value{
...
}
\description{
Check the use of dots
}
\keyword{internal}
