% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckDuplicateCellNames}
\alias{CheckDuplicateCellNames}
\title{Check a list of objects for duplicate cell names}
\usage{
CheckDuplicateCellNames(object.list, verbose = TRUE, stop = FALSE)
}
\arguments{
\item{object.list}{List of Seurat objects}

\item{verbose}{Print message about renaming}

\item{stop}{Error out if any duplicate names exist}
}
\value{
Returns list of objects with duplicate cells renamed to be unique
}
\description{
Check a list of objects for duplicate cell names
}
\keyword{internal}
