% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckFeaturesNames}
\alias{CheckFeaturesNames}
\title{Check features names format}
\usage{
CheckFeaturesNames(data)
}
\arguments{
\item{data}{a matrix input, rownames(data) are feature names}
}
\description{
Check features names format
}
