% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckGC}
\alias{CheckGC}
\title{Conditional Garbage Collection}
\usage{
CheckGC(option = "SeuratObject.memsafe")
}
\arguments{
\item{option}{...}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Call \code{gc} only when desired
}
\concept{utils}
