% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckLayersName}
\alias{CheckLayersName}
\title{Check layers names for the input list}
\usage{
CheckLayersName(matrix.list, layers.type = c("counts", "data"))
}
\arguments{
\item{matrix.list}{A list of matrices}

\item{layers.type}{layers type, such as counts or data}
}
\description{
Check layers names for the input list
}
\concept{utils}
