% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ClassKey}
\alias{ClassKey}
\title{Generate a Class Key}
\usage{
ClassKey(class, package = NULL)
}
\arguments{
\item{class}{Class name}

\item{package}{Optional name of package; by default, will search namespaces
of loaded packages to determine the providing package}
}
\value{
The class key (\dQuote{\code{package:class}})
}
\description{
Generate class keys for S4 classes. A class key follows the following
structure: \dQuote{\code{package:class}}
}
\examples{
ClassKey("Seurat")

}
