% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classpkg}
\alias{classpkg}
\alias{SwapClassPkg}
\alias{UpdateClassPkg}
\title{Update a Class's Package}
\usage{
SwapClassPkg(x, from = NULL, to = NULL)

UpdateClassPkg(object, from = NULL, to = NULL)
}
\arguments{
\item{x}{A list with an S4 class definition attribute}

\item{from}{A vector of one or more packages to limit conversion from}

\item{to}{A character naming the package to search for new class definitions;
defaults to the package of the function calling this function}

\item{object}{An S4 object}
}
\value{
\code{SwapClassPkg}: \code{x} with an updated S4 class
definition attribute

\code{UpdateClassPkg}: \code{object} with the updated
class definition
}
\description{
Swap packages for an object's class definition. As classes move between
packages, these functions rescope the namespace of the S4 class. This allows
objects to depend only on the new package for class definitions rather than
both the new and old packages
}
\section{S4 Class Definition Attributes}{

S4 classes are scoped to the package and class name. In order to properly
track which class a list is generated from in order to build a new one,
these function use an \code{\link[base:attr]{attribute}} to denote the
class name and package of origin. This attribute is stored as
\dQuote{classDef} and takes the form of \dQuote{\code{package:class}}.
}

\seealso{
\code{\link{s4list}}
}
\keyword{internal}
