% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{Collections}
\alias{Collections}
\title{Object Collections}
\usage{
Collections(object)
}
\arguments{
\item{object}{An S4 object}
}
\value{
A vector with the names of slots that are a list
}
\description{
Find the names of collections in an object
}
\examples{
\donttest{
SeuratObject:::Collections(pbmc_small)
}

}
\keyword{internal}
