% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Command}
\alias{Command}
\title{Get SeuratCommands}
\usage{
Command(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Either a SeuratCommand object or the requested paramter value
}
\description{
Pull information on previously run commands in the Seurat object.
}
