% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{CreateCentroids}
\alias{CreateCentroids}
\title{Create a \code{\link[SeuratObject:Centroids-class]{Centroids}} Objects}
\usage{
CreateCentroids(coords, nsides, radius, theta)
}
\arguments{
\item{coords}{The coordinates of cell/spot centroids}

\item{nsides}{The number of sides to represent cells/spots; pass
\code{\link[base]{Inf}} to plot as circles}

\item{radius}{Radius of shapes when plotting}

\item{theta}{Angle to adjust shapes when plotting}
}
\value{
A \code{\link[SeuratObject:Centroids-class]{Centroids}} object
}
\description{
Create a \code{\link[SeuratObject:Centroids-class]{Centroids}} Objects
}
