% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/molecules.R
\name{CreateMolecules}
\alias{CreateMolecules}
\alias{CreateMolecules.data.frame}
\alias{CreateMolecules.Molecules}
\alias{CreateMolecules.NULL}
\title{Create a \code{\link{Molecules}} Object}
\usage{
CreateMolecules(coords, ...)

\method{CreateMolecules}{data.frame}(coords, key = "", ...)

\method{CreateMolecules}{Molecules}(coords, ...)

\method{CreateMolecules}{`NULL`}(coords, ...)
}
\arguments{
\item{coords}{Spatial coordinates for molecules; should be a data frame
with three columns:
\itemize{
 \item \dQuote{\code{x}}: x-coordinates for each molecule
 \item \dQuote{\code{y}}: y-coordinates for each molecule
 \item \dQuote{\code{gene}}: gene name for each molecule
}}

\item{...}{Arguments passed to other methods}

\item{key}{A key to set for the molecules}
}
\value{
A \code{\link{Molecules}} object
}
\description{
Create a \code{\link{Molecules}} Object
}
