% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/segmentation.R
\name{CreateSegmentation}
\alias{CreateSegmentation}
\alias{CreateSegmentation.data.frame}
\alias{CreateSegmentation.Segmentation}
\alias{CreateSegmentation.sf}
\title{Create a \code{\link[SeuratObject:Segmentation-class]{Segmentation}} Objects}
\usage{
CreateSegmentation(coords)

\method{CreateSegmentation}{data.frame}(coords)

\method{CreateSegmentation}{Segmentation}(coords)

\method{CreateSegmentation}{sf}(coords)
}
\arguments{
\item{coords}{The coordinates of cell segmentations}
}
\value{
A \code{\link[SeuratObject:Segmentation-class]{Segmentation}} object
}
\description{
Create a \code{\link[SeuratObject:Segmentation-class]{Segmentation}} Objects
}
\concept{spatial}
