% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/fov.R
\name{Crop}
\alias{Crop}
\alias{Crop.FOV}
\title{Crop Coordinates}
\usage{
Crop(
  object,
  x = NULL,
  y = NULL,
  coords = c("plot", "tissue"),
  image.scale = "lowres",
  ...
)

\method{Crop}{FOV}(object, x = NULL, y = NULL, coords = c("plot", "tissue"), ...)
}
\arguments{
\item{object}{An object}

\item{x, y}{Range to crop x/y limits to; if \code{NULL}, uses full range of
\code{x}/\code{y}}

\item{coords}{Coordinate system to execute crop; choose from:
\itemize{
 \item \dQuote{\code{plot}}: Coordinates as shown when plotting
 \item \dQuote{\code{tissue}}: Coordinates from
  \code{\link{GetTissueCoordinates}}
}}

\item{image.scale}{Image scale to use when cropping images, 
should follow the scale of the image present in the FOV object. 
Currently only used when cropping FOVs that inherit from VisiumV2}

\item{...}{Arguments passed to other methods}
}
\value{
\code{object} cropped to the region specified by \code{x}
and \code{y}
}
\description{
Crop Coordinates
}
\concept{spatial}
