% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{DefaultAssay-StdAssay}
\alias{DefaultAssay-StdAssay}
\alias{DefaultAssay.StdAssay}
\alias{DefaultAssay<-.StdAssay}
\title{Default Assay}
\usage{
\method{DefaultAssay}{StdAssay}(object, ...)

\method{DefaultAssay}{StdAssay}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Name of assay to set as default}
}
\value{
\code{DefaultAssay}: The name of the default assay

\code{DefaultAssay<-}: An object with the default assay updated
}
\description{
Get and set the default assay
}
\keyword{internal}
