% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DefaultDimReduc}
\alias{DefaultDimReduc}
\title{Find the default \code{\link{DimReduc}}}
\usage{
DefaultDimReduc(object, assay = NULL)
}
\arguments{
\item{object}{A \code{\link{Seurat}} object}

\item{assay}{Name of assay to use; defaults to the default assay of the object}
}
\value{
The default \code{\link{DimReduc}}, if possible
}
\description{
Searches for \code{\link{DimReduc}s} matching \dQuote{umap}, \dQuote{tsne},
or \dQuote{pca}, case-insensitive, and in that order. Priority given to
\code{\link{DimReduc}s} matching the \code{DefaultAssay} or assay specified
(eg. \dQuote{pca} for the default assay weights higher than \dQuote{umap}
for a non-default assay)
}
\examples{
DefaultDimReduc(pbmc_small)

}
\concept{utils}
