% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/seurat.R
\name{DefaultFOV}
\alias{DefaultFOV}
\alias{DefaultFOV<-}
\alias{DefaultFOV.Seurat}
\alias{DefaultFOV<-.Seurat}
\title{Get and Set the Default FOV}
\usage{
DefaultFOV(object, ...)

DefaultFOV(object, ...) <- value

\method{DefaultFOV}{Seurat}(object, assay = NULL, ...)

\method{DefaultFOV}{Seurat}(object, assay = NA, ...) <- value
}
\arguments{
\item{object}{A \code{\link{Seurat}} Object}

\item{...}{Arguments passed to other methods}

\item{value}{The name of the \code{\link{FOV}} to set as the default}

\item{assay}{Name of assay to get or set default \code{\link{FOV}} for;
pass \code{NA} to get or set the global default \code{\link{FOV}}}
}
\value{
\code{DefaultFOV}: The name of the default \code{\link{FOV}}

\code{DefaultFOV<-}: \code{object} with the default FOV set
to \code{value}
}
\description{
Get and Set the Default FOV
}
