% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{DefaultImage}
\alias{DefaultImage}
\title{Get the default image of an object}
\usage{
DefaultImage(object)
}
\arguments{
\item{object}{A \code{\link{Seurat}} object}
}
\value{
The name of the default image
}
\description{
Attempts to find all images associated with the default assay of the object.
If none present, finds all images present in the object. Returns the name of
the first image
}
\keyword{internal}
