% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{DefaultLayer-StdAssay}
\alias{DefaultLayer-StdAssay}
\alias{DefaultLayer.StdAssay}
\alias{DefaultLayer<-.StdAssay}
\title{Default Layer}
\usage{
\method{DefaultLayer}{StdAssay}(object, ...)

\method{DefaultLayer}{StdAssay}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Name of layer to set as default}
}
\value{
\code{DefaultLayer}: The name of the default layer

\code{DefaultLayer<-}: An object with the default layer updated
}
\description{
Get and set the default layer
}
\keyword{internal}
