% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{dim.Assay}
\alias{dim.Assay}
\title{Feature and Cell Numbers}
\usage{
\method{dim}{Assay}(x)
}
\arguments{
\item{x}{An \code{\link{Assay}} object}
}
\value{
A two-length numeric vector with the total number of
features and cells in \code{x}
}
\description{
Feature and Cell Numbers
}
\examples{
rna <- pbmc_small[["RNA"]]
dim(rna)

}
\seealso{
v3 Assay object, validity, and interaction methods:
\code{\link{$.Assay}()},
\code{\link{Assay-class}},
\code{\link{Assay-validity}},
\code{\link{CreateAssayObject}()},
\code{\link{[.Assay}()},
\code{\link{[[.Assay}()},
\code{\link{dimnames.Assay}()},
\code{\link{merge.Assay}()},
\code{\link{split.Assay}()},
\code{\link{subset.Assay}()}
}
\concept{assay}
