% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{dim.Assay5}
\alias{dim.Assay5}
\title{Feature and Cell Numbers}
\usage{
\method{dim}{Assay5}(x)
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}
}
\value{
A two-length numeric vector with the total number of
features and cells in \code{x}
}
\description{
Feature and Cell Numbers
}
\seealso{
v5 Assay object, validity, and interaction methods:
\code{\link{$.Assay5}()},
\code{\link{Assay5-class}},
\code{\link{Assay5-validity}},
\code{\link{[.Assay5}()},
\code{\link{[[.Assay5}()},
\code{\link{dimnames.Assay5}()},
\code{\link{merge.Assay5}()},
\code{\link{split.Assay5}()},
\code{\link{subset.Assay5}()}
}
\concept{assay5}
