% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{dim.Seurat}
\alias{dim.Seurat}
\title{Feature and Cell Numbers}
\usage{
\method{dim}{Seurat}(x)
}
\arguments{
\item{x}{A \code{\link{Seurat}} object}
}
\value{
A two-length numeric vector with the total number of
features and cells in \code{x}
}
\description{
Feature and Cell Numbers
}
\examples{
# Get the number of features in an object
nrow(pbmc_small)

# Get the number of cells in an object
ncol(pbmc_small)

}
\seealso{
Seurat object, validity, and interaction methods
\code{\link{$.Seurat}()},
\code{\link{Seurat-class}},
\code{\link{Seurat-validity}},
\code{\link{[[.Seurat}()},
\code{\link{[[<-,Seurat,NULL}},
\code{\link{[[<-,Seurat}},
\code{\link{dimnames.Seurat}()},
\code{\link{merge.Seurat}()},
\code{\link{names.Seurat}()},
\code{\link{subset.Seurat}()}
}
\concept{seurat}
