% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{dim.StdAssay}
\alias{dim.StdAssay}
\title{Feature and Cell Numbers}
\usage{
\method{dim}{StdAssay}(x)
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}
}
\value{
A two-length numeric vector with the total number of
features and cells in \code{x}
}
\description{
Feature and Cell Numbers
}
\seealso{
v5 Standard Assay object, validity, and interaction methods
\code{\link{$.StdAssay}()},
\code{\link{.DollarNames.StdAssay}()},
\code{\link{StdAssay-class}},
\code{\link{StdAssay-validity}},
\code{\link{[.StdAssay}()},
\code{\link{[[.StdAssay}()},
\code{\link{dimnames.StdAssay}()},
\code{\link{show,StdAssay-method}},
\code{\link{split.StdAssay}()},
\code{\link{subset.StdAssay}()}
}
\concept{stdassay}
\keyword{internal}
