% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{dimnames.Assay5}
\alias{dimnames.Assay5}
\alias{dimnames<-.Assay5}
\title{Assay-Level Feature and Cell Names}
\usage{
\method{dimnames}{Assay5}(x)

\method{dimnames}{Assay5}(x) <- value
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{value}{A two-length list with updated feature and/or cells names}
}
\value{
\code{dimnames}: A two-length list with the following values:
\itemize{
 \item A character vector will all features in \code{x}
 \item A character vector will all cells in \code{x}
}

\code{dimnames<-}: \code{x} with the feature and/or cell
names updated to \code{value}
}
\description{
Get and set feature and cell names in v5 Assays
}
\seealso{
Other assay5: 
\code{\link{Assay5-class}},
\code{\link{Assay5-validity}},
\code{\link{[.Assay5}()},
\code{\link{[[.Assay5}()},
\code{\link{dim.Assay5}()},
\code{\link{merge.Assay5}()},
\code{\link{subset.Assay5}()}

Other dimnames: 
\code{\link{Cells}()},
\code{\link{dimnames.Assay}()}
}
\concept{assay5}
\concept{dimnames}
