% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.AddMetaData}
\alias{.AddMetaData}
\title{Object Metadata}
\usage{
.AddMetaData(object, metadata, col.name = NULL)
}
\arguments{
\item{object}{An object}

\item{metadata}{A vector, list, or data.frame with metadata to add}

\item{col.name}{A name for meta data if not a named list or data.frame}
}
\value{
object with metadata added
}
\description{
Internal \code{\link{AddMetaData}} defintion
}
