% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/default.R, R/utils.R
\name{.AssayClass}
\alias{.AssayClass}
\alias{.AssayClass.StdAssay}
\alias{.AssayClass.default}
\title{Assay Class Label}
\usage{
.AssayClass(object)

\method{.AssayClass}{StdAssay}(object)

\method{.AssayClass}{default}(object)
}
\arguments{
\item{object}{A \code{\link{StdAssay}} object}
}
\value{
The assay class label for \code{object}
}
\description{
Assay Class Label
}
\keyword{internal}
