% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.BPMatrixMode}
\alias{.BPMatrixMode}
\title{\pkg{BPCells} Matrix Mode}
\usage{
.BPMatrixMode(object, simplify = FALSE)
}
\arguments{
\item{object}{An \code{IterableMatrix}}

\item{simplify}{Return \dQuote{\code{disk}} for on-disk matrices}
}
\value{
One of the following, depending on the mode of \code{object}:
\itemize{
 \item \dQuote{\code{memory}}
 \item \dQuote{\code{file}}
 \item \dQuote{\code{directory}}
}
If \code{simplify} is \code{TRUE}, returns \dQuote{\code{disk}} instead of
\dQuote{\code{file}} or \dQuote{\code{directory}}
}
\description{
Get the mode (on-disk, in-memory) of an \code{IterableMatrix} object
from \pkg{BPCells}
}
\keyword{internal}
