% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{.CalcN}
\alias{.CalcN}
\title{Calculate nCount and nFeature}
\usage{
.CalcN(object, ...)
}
\arguments{
\item{object}{An assay-like object}

\item{...}{Arguments passed to other methods}
}
\value{
A named list with ...
}
\description{
Calculate nCount and nFeature
}
\examples{
calcn <- .CalcN(pbmc_small[["RNA"]])
head(as.data.frame(calcn))

}
\keyword{internal}
