% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{.CheckFmargin}
\alias{.CheckFmargin}
\title{Check Feature Margin}
\usage{
.CheckFmargin(fmargin)
}
\arguments{
\item{fmargin}{Either \code{1} or \code{2}}
}
\value{
\code{fmargin}
}
\description{
Check Feature Margin
}
\examples{
.CheckFmargin(1L)
.CheckFmargin(2.3)

# Error if `fmargin` is outside of [1:2]
if (FALSE) {
  .CheckFmargin(3L)
}

}
\keyword{internal}
