% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.Collections}
\alias{.Collections}
\title{Identify Object Collections}
\usage{
.Collections(object, exclude = character(length = 0L), ...)
}
\arguments{
\item{object}{An \link[methods:Classes_Details]{S4} object}

\item{exclude}{A character vector of slot names to exclude}

\item{...}{Arguments passed to \code{\link{IsNamedList}}}
}
\value{
A character vector of names of collection slots
}
\description{
Find all collection (named lists) slots in an S4 object
}
\examples{
.Collections(pbmc_small)

}
\seealso{
Other subobjects: 
\code{\link{.FilterObjects}()},
\code{\link{.FindObject}()}
}
\concept{subobjects}
\concept{utils}
\keyword{internal}
