% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.Contains}
\alias{.Contains}
\title{Get Parent S4 Classes}
\usage{
.Contains(object)
}
\arguments{
\item{object}{An \link[methods:Classes_Details]{S4} object}
}
\value{
A vector of class names that \code{object} inherits from
}
\description{
Get Parent S4 Classes
}
\examples{
.Contains(pbmc_small)

}
\concept{utils}
\keyword{internal}
