% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.DefaultFOV}
\alias{.DefaultFOV}
\title{Find the Default FOV}
\usage{
.DefaultFOV(object, assay = NULL)
}
\arguments{
\item{object}{A \code{{Seurat}} object}
}
\value{
...
}
\description{
Attempts to find the \dQuote{default} FOV using the revamped
spatial framework
}
\keyword{internal}
