% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.DefaultSpatialCoords}
\alias{.DefaultSpatialCoords}
\title{Find the Default SpatialCoords Image}
\usage{
.DefaultSpatialCoords(object, assay = NULL)
}
\arguments{
\item{object}{A \code{{Seurat}} object}
}
\value{
...
}
\description{
Attempts to find the \dQuote{default} spatial image using the revamped
spatial framework
}
\keyword{internal}
