% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{.DollarNames.Assay}
\alias{.DollarNames.Assay}
\title{Dollar-sign Autocompletion}
\usage{
\method{.DollarNames}{Assay}(x, pattern = "")
}
\arguments{
\item{x}{An \code{\link{Assay}} object}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }
}
\value{
The layer name matches for \code{pattern}
}
\description{
Autocompletion for \code{$} access on an
\code{\link{Assay}} object
}
\seealso{
\code{\link[utils:.DollarNames]{utils::.DollarNames}}
}
\concept{assay}
\keyword{internal}
