% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{.DollarNames.Seurat}
\alias{.DollarNames.Seurat}
\title{Dollar-sign Autocompletion}
\usage{
\method{.DollarNames}{Seurat}(x, pattern = "")
}
\arguments{
\item{x}{A \code{\link{Seurat}} object}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }
}
\value{
The meta data matches for \code{pattern}
}
\description{
Autocompletion for \code{$} access on a \code{\link{Seurat}} object
}
\seealso{
\code{\link[utils:.DollarNames]{utils::.DollarNames}}
}
\concept{seurat}
\keyword{internal}
