% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{.DollarNames.SeuratCommand}
\alias{.DollarNames.SeuratCommand}
\title{Dollar-sign Autocompletion}
\usage{
\method{.DollarNames}{SeuratCommand}(x, pattern = "")
}
\arguments{
\item{x}{A \code{\link{SeuratCommand}} object}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }
}
\value{
The parameter name matches for \code{pattern}
}
\description{
Autocompletion for \code{$} access on a
\code{\link{SeuratCommand}} object
}
\seealso{
Command log object and interaction methods
\code{\link{$.SeuratCommand}()},
\code{\link{LogSeuratCommand}()},
\code{\link{SeuratCommand-class}},
\code{\link{[.SeuratCommand}()},
\code{\link{as.list.SeuratCommand}()}
}
\concept{command}
