% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{.DollarNames.StdAssay}
\alias{.DollarNames.StdAssay}
\title{Dollar-sign Autocompletion}
\usage{
\method{.DollarNames}{StdAssay}(x, pattern = "")
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }
}
\value{
The layer name matches for \code{pattern}
}
\description{
Autocompletion for \code{$} access on an \code{\link{Assay5}} object
}
\seealso{
v5 Standard Assay object, validity, and interaction methods
\code{\link{$.StdAssay}()},
\code{\link{StdAssay-class}},
\code{\link{StdAssay-validity}},
\code{\link{[.StdAssay}()},
\code{\link{[[.StdAssay}()},
\code{\link{dim.StdAssay}()},
\code{\link{dimnames.StdAssay}()},
\code{\link{show,StdAssay-method}},
\code{\link{split.StdAssay}()},
\code{\link{subset.StdAssay}()}
}
\concept{stdassay}
\keyword{internal}
