% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{.FilePath}
\alias{.FilePath}
\alias{.FilePath.default}
\alias{.FilePath.DelayedMatrix}
\title{Find a File Path}
\usage{
.FilePath(x)

\method{.FilePath}{default}(x)

\method{.FilePath}{DelayedMatrix}(x)
}
\arguments{
\item{x}{A file-backed object}
}
\value{
The path to the file that backs \code{x}; if \code{x} is not a
file-backed object, returns \code{NULL}
}
\description{
Find a File Path
}
\keyword{internal}
