% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{.GetConsensusFeatures}
\alias{.GetConsensusFeatures}
\title{Return the n most frequently observed features in `features_by_layer`, where
n = `nfeatures`. If two features are observed at the same frequency,
thee median index they appear in will be used to resolve their order. If
`nfeatures` is not specified, all features found in `common_features` are
returned.}
\usage{
.GetConsensusFeatures(features_by_layer, common_features, nfeatures = NULL)
}
\description{
Return the n most frequently observed features in `features_by_layer`, where
n = `nfeatures`. If two features are observed at the same frequency,
thee median index they appear in will be used to resolve their order. If
`nfeatures` is not specified, all features found in `common_features` are
returned.
}
