% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.GetMethod}
\alias{.GetMethod}
\title{Get a Method}
\usage{
.GetMethod(fxn, cls)
}
\arguments{
\item{fxn}{Name of a function as a character}

\item{cls}{...}
}
\value{
...
}
\description{
Get a Method
}
\examples{
.GetMethod('t', 'Matrix')
.GetMethod('t', 'data.frame')

}
\keyword{internal}
