% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.IsFutureSeurat}
\alias{.IsFutureSeurat}
\alias{IsFutureSeurat}
\title{Test Future Compatibility with \pkg{Seurat}}
\usage{
.IsFutureSeurat(version, lib.loc = NULL)
}
\arguments{
\item{version}{A version string or object of class
\code{\link{package_version}}}

\item{lib.loc}{a character vector of directory names of \R libraries,
    or \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.  If the default is used, the loaded
    packages and namespaces are searched before the libraries.}
}
\value{
\code{TRUE} if \pkg{SeuratObject} and/or \pkg{Seurat}
}
\description{
Check to see if \pkg{SeuratObject} and/or \pkg{Seurat} are at least a
specific version or if they're configured to act as if they're a
specific version (see details below). This allows testing compatibility with
future requirements for both \pkg{SeuratObject} and \pkg{Seurat}
}
\details{
Blah blah blah
}
