% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keymixin.R
\name{.KeyPattern}
\alias{.KeyPattern}
\title{Regex Pattern for Keys}
\usage{
.KeyPattern()
}
\value{
Returns the regex pattern for keys
(\dQuote{\Sexpr[stage=build]{SeuratObject:::.KeyPattern()}})
}
\description{
Regex Pattern for Keys
}
\seealso{

\code{\link{.RandomKey}()},
\code{\link{Key-validity}},
\code{\link{KeyMixin-class}}
}
\concept{key}
\keyword{internal}
