% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{.MARGIN}
\alias{.MARGIN}
\title{Get the Margin of an Object}
\usage{
.MARGIN(x, ...)
}
\arguments{
\item{x}{An object}
}
\value{
The margin, eg. \code{1} for rows or \code{2} for columns
}
\description{
Get the Margin of an Object
}
\keyword{internal}
