% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.Misc}
\alias{.Misc}
\alias{.Misc<-}
\title{Miscellaneous Data}
\usage{
.Misc(object, slot = NULL, ...)

.Misc(object, slot, ...) <- value
}
\arguments{
\item{object}{An object}

\item{slot}{Name of miscellaneous data to get or set}

\item{...}{Arguments passed to other methods}

\item{value}{Data to add}
}
\value{
\code{.Misc}: If \code{slot} is \code{NULL}, all miscellaneous
data, otherwise the miscellaneous data for \code{slot}

\code{.Misc<-}: \code{object} with \code{value} added to the
miscellaneous data slot \code{slot}
}
\description{
Internal functions for getting and setting miscellaneous data
}
\keyword{internal}
