% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keymixin.R
\name{.RandomKey}
\alias{.RandomKey}
\title{Generate a Random Key}
\usage{
.RandomKey(length = 7L, ...)
}
\arguments{
\item{length}{How long should the name be}

\item{...}{Extra parameters passed to \code{\link[base]{sample}}}
}
\value{
Returns a valid key
}
\description{
Generate a Random Key
}
\examples{
set.seed(42L)
.RandomKey()

}
\seealso{

\code{\link{.KeyPattern}()},
\code{\link{Key-validity}},
\code{\link{KeyMixin-class}}
}
\concept{key}
\keyword{internal}
