% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{.SelectFeatures}
\alias{.SelectFeatures}
\alias{.SelectFeatures.list}
\title{Combine and Select Features}
\usage{
.SelectFeatures(object, ...)

\method{.SelectFeatures}{list}(object, all.features = NULL, nfeatures = Inf, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
A vector of features selected
}
\description{
Combine and Select Features
}
\keyword{internal}
