% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{.SparseSlots}
\alias{.SparseSlots}
\alias{.SparseSlots.CsparseMatrix}
\alias{.SparseSlots.RsparseMatrix}
\alias{.SparseSlots.spam}
\title{Identify Sparse Slots}
\usage{
.SparseSlots(x, type = c("pointers", "indices", "entries"))

\method{.SparseSlots}{CsparseMatrix}(x, type = c("pointers", "entries", "indices"))

\method{.SparseSlots}{RsparseMatrix}(x, type = c("pointers", "indices", "entries"))

\method{.SparseSlots}{spam}(x, type = c("pointers", "indices", "entries"))
}
\arguments{
\item{x}{A sparse matrix}

\item{type}{...}
}
\value{
...
}
\description{
Identify Sparse Slots
}
\seealso{
Other sparse: 
\code{\link{IsSparse}()},
\code{\link{RegisterSparseMatrix}()}
}
\concept{sparse}
\keyword{internal}
