% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.Subobjects}
\alias{.Subobjects}
\title{Get the Subobject Names}
\usage{
.Subobjects(object, exclude = c("misc", "tools"), collapse = TRUE, ...)
}
\arguments{
\item{object}{An \link[methods:Classes_Details]{S4} object}

\item{exclude}{A character vector of slot names to exclude}

\item{collapse}{Collapse the list into a vector}

\item{...}{Arguments passed to \code{\link{IsNamedList}}}
}
\value{
If \code{collapse = TRUE}, then a vector with the names of all
subobjects; otherwise, a named list where the names are the names of the
collections and the values are the names of subobjects within the collection
}
\description{
Get the Subobject Names
}
\examples{
.Subobjects(pbmc_small)

}
\seealso{

\code{\link{.Collections}()},
\code{\link{.FilterObjects}()},
\code{\link{.FindObject}()}
}
\concept{subobjects}
\keyword{internal}
\keyword{utils}
