% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimreduc.R
\name{Embeddings}
\alias{Embeddings}
\alias{Embeddings.DimReduc}
\title{Get cell embeddings}
\usage{
Embeddings(object, ...)

\method{Embeddings}{DimReduc}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get cell embeddings
}
\examples{
# Get the embeddings directly from a DimReduc object
Embeddings(object = pbmc_small[["pca"]])[1:5, 1:5]

}
