% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EmptyDF}
\alias{EmptyDF}
\title{Empty Data Frames}
\usage{
EmptyDF(n)
}
\arguments{
\item{n}{Number of rows for the data frame}
}
\value{
A \link[base:data.frame]{data frame} with \code{n} rows and
zero columns
}
\description{
Create an empty \link[base:data.frame]{data frame} with no row names and
zero columns
}
\keyword{internal}
