% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{FilterObjects}
\alias{FilterObjects}
\title{Find Sub-objects of a Certain Class}
\usage{
FilterObjects(object, classes.keep = c("Assay", "StdAssay", "DimReduc"))
}
\arguments{
\item{object}{A \code{\link{Seurat}} object}

\item{classes.keep}{A vector of names of classes to get}
}
\value{
A vector with the names of objects within the \code{Seurat} object
that are of class \code{classes.keep}
}
\description{
Get the names of objects within a \code{Seurat} object that are of a
certain class
}
\examples{
FilterObjects(pbmc_small)

}
\concept{utils}
