% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fov.R
\name{FOV-validity}
\alias{FOV-validity}
\title{FOV Validity}
\description{
Validation of \code{FOV} objects is handled by
\code{\link[methods]{validObject}}
}
\section{Boundary Validation}{

blah
}

\section{Molecule Validation}{

blah
}

\seealso{
\code{\link[methods]{validObject}}
}
\concept{fov}
