% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{GetAssay}
\alias{GetAssay}
\title{Get an Assay object from a given Seurat object.}
\usage{
GetAssay(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns an Assay object
}
\description{
Get an Assay object from a given Seurat object.
}
