% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R
\name{GetAssayData}
\alias{GetAssayData}
\alias{GetAssayData.Assay}
\title{General accessor function for the Assay class}
\usage{
GetAssayData(object, ...)

\method{GetAssayData}{Assay}(object, slot = c("data", "scale.data", "counts"), ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{slot}{Specific information to pull (i.e. counts, data, scale.data, ...)}
}
\value{
Returns info from requested slot
}
\description{
This function can be used to pull information from any of the slots in the
Assay class. For example, pull one of the data matrices("counts", "data", or
"scale.data").
}
\examples{
# Get the data directly from an Assay object
GetAssayData(object = pbmc_small[["RNA"]], slot = "data")[1:5,1:5]

}
\concept{data-access}
