% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{GetPolygonCoordinates}
\alias{GetPolygonCoordinates}
\title{Get coordinates for polygon plotting from an \code{sf} object}
\usage{
GetPolygonCoordinates(sf_data)
}
\arguments{
\item{sf_data}{A \code{sf} object}
}
\value{
A data.frame with columns x, y, and cell_id, such that each row
represents a single vertex of the segmentation.
}
\description{
Extract coordinates from an \code{sf} object in a format suitable
for use with \code{ggplot2::geom_polygon()}.
}
