% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R, R/seurat.R
\name{GetTissueCoordinates}
\alias{GetTissueCoordinates}
\alias{GetTissueCoordinates.SlideSeq}
\alias{GetTissueCoordinates.STARmap}
\alias{GetTissueCoordinates.VisiumV1}
\alias{GetTissueCoordinates.Seurat}
\title{Get tissue coordinates}
\usage{
GetTissueCoordinates(object, ...)

\method{GetTissueCoordinates}{SlideSeq}(object, ...)

\method{GetTissueCoordinates}{STARmap}(object, qhulls = FALSE, ...)

\method{GetTissueCoordinates}{VisiumV1}(
  object,
  scale = "lowres",
  cols = c("imagerow", "imagecol"),
  ...
)

\method{GetTissueCoordinates}{Seurat}(object, image = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{qhulls}{return qhulls instead of centroids}

\item{scale}{A factor to scale the coordinates by; choose from: 'tissue',
'fiducial', 'hires', 'lowres', or \code{NULL} for no scaling}

\item{cols}{Columns of tissue coordinates data.frame to pull}

\item{image}{Name of \code{SpatialImage} object to get coordinates for; if
\code{NULL}, will attempt to select an image automatically}
}
\value{
A data.frame with tissue coordinates
}
\description{
Get tissue coordinates
}
\concept{data-access}
