% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{class}
\name{Graph-class}
\alias{Graph-class}
\alias{Graph}
\title{The Graph Class}
\description{
The Graph class inherits from \code{\link[Matrix:sparseMatrix]{dgCMatrix}}.
We do this to enable future expandability of graphs.
}
\section{Slots}{

\describe{
\item{\code{assay.used}}{Optional name of assay used to generate \code{Graph} object}
}}

\seealso{
\code{\link[Matrix]{dgCMatrix-class}}

Other graph: 
\code{\link{as.Graph}()}
}
\concept{graph}
