% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsNullPtr}
\alias{IsNullPtr}
\title{Test Null Pointers}
\usage{
IsNullPtr(x)
}
\arguments{
\item{x}{An \link[methods:externalptr-class]{external pointer} object}
}
\value{
\code{TRUE} if \code{x} is a null pointer, otherwise \code{FALSE}
}
\description{
Check to see if a C++ pointer is a null pointer on the compiled side
}
\references{
\url{https://stackoverflow.com/questions/26666614/how-do-i-check-if-an-externalptr-is-null-from-within-r}
}
\keyword{internal}
