% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{IsSCT}
\alias{IsSCT}
\title{Check whether an assay has been processed by sctransform}
\usage{
IsSCT(assay)
}
\arguments{
\item{assay}{assay to check}
}
\value{
Boolean
}
\description{
Check whether an assay has been processed by sctransform
}
\keyword{internal}
