% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{IsSparse}
\alias{IsSparse}
\title{Is a Matrix Sparse}
\usage{
IsSparse(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
...
}
\description{
Is a Matrix Sparse
}
\examples{
IsSparse(matrix())
IsSparse(LayerData(pbmc_small, "counts"))

}
\seealso{

\code{\link{.SparseSlots}()},
\code{\link{RegisterSparseMatrix}()}
}
\concept{sparse}
\keyword{internal}
