% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keymixin.R
\name{Key-validity}
\alias{Key-validity}
\title{Key Validity}
\description{
Validation of \code{\link{KeyMixin}} objects is handled by
\code{\link[methods]{validObject}}
}
\section{Key Validation}{

Keys must be a one-length character vector; a key must be composed of one
of the following:
\itemize{
 \item An empty string (eg. \dQuote{\code{''}}) where \code{nzchar() == 0}
 \item An string composed of one or more alphanumeric values
 (both lower- and upper-case) that ends with an underscore
 (\dQuote{\code{_}}); the first character must be a letter
}
Keys that are not empty strings are validated with the regex
\dQuote{\code{\Sexpr[stage=build]{SeuratObject:::.KeyPattern()}}}
}

\seealso{

\code{\link{.KeyPattern}()},
\code{\link{.RandomKey}()},
\code{\link{KeyMixin-class}}
}
\concept{key}
\keyword{internal}
