% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay5.R
\name{LayerData}
\alias{LayerData}
\alias{LayerData<-}
\alias{Layers}
\alias{Layers.StdAssay}
\title{Query and Manipulate Assay Layers}
\usage{
LayerData(object, layer = "data", ...)

LayerData(object, layer, ...) <- value

Layers(object, ...)

\method{Layers}{StdAssay}(object, data = TRUE, ...)
}
\arguments{
\item{object}{An object}

\item{layer}{Name of layer to fetch or set}

\item{...}{Arguments passed to other methods}

\item{value}{...}

\item{data}{Include \dQuote{data} as a layer}
}
\value{
...
}
\description{
Query and Manipulate Assay Layers
}
\concept{data-access}
