% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimreduc.R
\name{Loadings}
\alias{Loadings}
\alias{Loadings<-}
\alias{Loadings.DimReduc}
\alias{Loadings<-.DimReduc}
\title{Get and set feature loadings}
\usage{
Loadings(object, ...)

Loadings(object, ...) <- value

\method{Loadings}{DimReduc}(object, projected = FALSE, ...)

\method{Loadings}{DimReduc}(object, projected = TRUE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Feature loadings to add}

\item{projected}{Pull the projected feature loadings?}
}
\description{
Get and set feature loadings
}
\examples{
# Get the feature loadings for a given DimReduc
Loadings(object = pbmc_small[["pca"]])[1:5,1:5]

# Set the feature loadings for a given DimReduc
new.loadings <- Loadings(object = pbmc_small[["pca"]])
new.loadings <- new.loadings + 0.01
Loadings(object = pbmc_small[["pca"]]) <- new.loadings

}
