% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{LogSeuratCommand}
\alias{LogSeuratCommand}
\title{Log a command}
\usage{
LogSeuratCommand(object, return.command = FALSE)
}
\arguments{
\item{object}{Name of Seurat object}

\item{return.command}{Return a \code{\link{SeuratCommand}} object instead}
}
\value{
If \code{return.command}, returns a \code{\link{SeuratCommand}}
object; otherwise, returns the Seurat object with command stored
}
\description{
Logs command run, storing the name, timestamp, and argument list. Stores in
the Seurat object
}
\seealso{
\code{\link{Command}}

Command log object and interaction methods
\code{\link{$.SeuratCommand}()},
\code{\link{.DollarNames.SeuratCommand}()},
\code{\link{SeuratCommand-class}},
\code{\link{[.SeuratCommand}()},
\code{\link{as.list.SeuratCommand}()}
}
\concept{command}
