% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R
\name{merge.DimReduc}
\alias{merge.DimReduc}
\title{Merge Dimensional Reductions}
\usage{
\method{merge}{DimReduc}(x = NULL, y = NULL, add.cell.ids = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{DimReduc}} object}

\item{y}{One or more \code{\link{DimReduc}} objects}

\item{add.cell.ids}{A character vector equal to the number of objects
provided to append to all cell names; if \code{TRUE}, uses \code{labels} as
\code{add.cell.ids}}

\item{...}{Ignored}
}
\value{
A new \code{DimReduc} object with data merged from \code{c(x, y)}
}
\description{
Merge two or more \link[=DimReduc]{dimensional reduction} together
}
\seealso{
Dimensional reduction object, validity, and interaction methods
\code{\link{CreateDimReducObject}()},
\code{\link{DimReduc-class}},
\code{\link{DimReduc-validity}},
\code{\link{[.DimReduc}()},
\code{\link{[[.DimReduc}()},
\code{\link{dim.DimReduc}()},
\code{\link{print.DimReduc}()},
\code{\link{subset.DimReduc}()}
}
\concept{dimreduc}
