% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{Misc-StdAssay}
\alias{Misc-StdAssay}
\alias{Misc.StdAssay}
\alias{Misc<-.StdAssay}
\title{Get and set miscellaneous data}
\usage{
\method{Misc}{StdAssay}(object, slot = NULL, ...)

\method{Misc}{StdAssay}(object, slot, ...) <- value
}
\arguments{
\item{object}{An object}

\item{slot}{Name of specific bit of meta data to pull}

\item{...}{Arguments passed to other methods}

\item{value}{Data to add}
}
\value{
Miscellaneous data

An object with miscellaneous data added
}
\description{
Get and set miscellaneous data
}
\keyword{internal}
