% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R, R/dimreduc.R
\name{Misc}
\alias{Misc}
\alias{Misc<-}
\alias{Misc.Assay}
\alias{Misc<-.Assay}
\alias{Misc.DimReduc}
\alias{Misc<-.DimReduc}
\title{Get and set miscellaneous data}
\usage{
Misc(object, ...)

Misc(object, ...) <- value

\method{Misc}{Assay}(object, slot = NULL, ...)

\method{Misc}{Assay}(object, slot, ...) <- value

\method{Misc}{DimReduc}(object, slot = NULL, ...)

\method{Misc}{DimReduc}(object, slot, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Data to add}

\item{slot}{Name of specific bit of meta data to pull}
}
\value{
Miscellaneous data

An object with miscellaneous data added
}
\description{
Get and set miscellaneous data
}
