% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecules.R
\docType{class}
\name{Molecules-class}
\alias{Molecules-class}
\title{The Spatial Molecules Class}
\description{
The Spatial Molecules Class
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list of \code{\link[sp]{SpatialPoints}} objects}

\item{\code{key}}{The key for the \code{Molecules}}
}}

\seealso{
\code{Molecules} methods: \code{\link{Molecules-methods}}

Segmentation layer classes:
\code{\link{Centroids-class}},
\code{\link{Centroids-methods}},
\code{\link{Molecules-methods}},
\code{\link{Segmentation-class}},
\code{\link{Segmentation-methods}},
\code{\link{Segmentation-validity}}
}
\concept{segmentation}
