% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{Neighbor-methods}
\alias{Neighbor-methods}
\alias{dim.Neighbor}
\alias{show,Neighbor-method}
\title{\code{Neighbor} Methods}
\usage{
\method{dim}{Neighbor}(x)

\S4method{show}{Neighbor}(object)
}
\arguments{
\item{x, object}{A \code{\link{Neighbor}} object}
}
\value{
\code{dim} Dimensions of the indices matrix

\code{show}: Prints summary to \code{\link[base]{stdout}} and
invisibly returns \code{NULL}
}
\description{
Methods for \code{\link{Neighbor}} objects for generics defined in
other packages
}
\section{Functions}{
\itemize{
\item \code{dim(Neighbor)}: Dimensions of the neighbor indices

\item \code{show(Neighbor)}: Overview of a \code{Neighbor} object

}}
\concept{neighbor}
