% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{NullImage}
\alias{NullImage}
\title{Return a null image}
\usage{
NullImage(mode = c("grob", "raster", "plotly", "raw"))
}
\arguments{
\item{mode}{How to return the image; should accept one of 'grob', 'raster',
'plotly', or 'raw'}
}
\value{
Varies by value of \code{mode}:
\describe{
 \item{\dQuote{grob}}{a \code{\link[grid]{nullGrob}}}
 \item{\dQuote{raster}}{an empty \code{\link[grDevices:as.raster]{raster}}}
 \item{\dQuote{plotly}}{a list with one named item: \code{value = FALSE}}
 \item{\dQuote{raw}}{returns \code{NULL}}
}
}
\description{
Return a null image
}
\keyword{internal}
