% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PackageCheck}
\alias{PackageCheck}
\title{Check the existence of a package}
\usage{
PackageCheck(..., error = TRUE)
}
\arguments{
\item{...}{Package names}

\item{error}{If true, throw an error if the package doesn't exist}
}
\value{
Invisibly returns boolean denoting if the package is installed
}
\description{
Check the existence of a package
}
\examples{
PackageCheck("SeuratObject", error = FALSE)

}
