% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Project}
\alias{Project}
\alias{Project<-}
\title{Get and set project information}
\usage{
Project(object, ...)

Project(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Project information to set}
}
\value{
Project information

An object with project information added
}
\description{
Get and set project information
}
