% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R
\name{Projected}
\alias{Projected}
\title{Check to see if projected loadings have been set}
\usage{
Projected(object)
}
\arguments{
\item{object}{a DimReduc object}
}
\value{
TRUE if projected loadings have been set, else FALSE
}
\description{
Check to see if projected loadings have been set
}
\keyword{internal}
