% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Radius}
\alias{Radius}
\title{Get the spot radius from an image}
\usage{
Radius(object, ...)
}
\arguments{
\item{object}{An image object}

\item{...}{Arguments passed to other methods}
}
\value{
The radius size
}
\description{
Get the spot radius from an image
}
\concept{spatialimage}
