% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RandomName}
\alias{RandomName}
\title{Generate a random name}
\usage{
RandomName(length = 5L, chars = letters, ...)
}
\arguments{
\item{length}{How long should the name be}

\item{chars}{A vector of 1-length characters to use to generate the name}

\item{...}{Extra parameters passed to \code{\link[base]{sample}}}
}
\value{
A character with \code{nchar == length} of randomly sampled letters
}
\description{
Make a name from randomly sampled characters, pasted together with no spaces
}
\examples{
set.seed(42L)
RandomName()
RandomName(7L, replace = TRUE)

}
\seealso{
\code{\link[base]{sample}}
}
\concept{utils}
