% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{RegisterSparseMatrix}
\alias{RegisterSparseMatrix}
\title{Register Sparse Matrix Classes}
\usage{
RegisterSparseMatrix(class, package = NULL)
}
\arguments{
\item{class}{Class name}

\item{package}{Optional name of package; by default, will search namespaces
of loaded packages to determine the providing package}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Register Sparse Matrix Classes
}
\seealso{
Other sparse: 
\code{\link{.SparseSlots}()},
\code{\link{IsSparse}()}
}
\concept{sparse}
\keyword{internal}
