% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{RenameCells-StdAssay}
\alias{RenameCells-StdAssay}
\alias{RenameCells.StdAssay}
\title{Rename cells}
\usage{
\method{RenameCells}{StdAssay}(object, new.names = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{new.names}{vector of new cell names}

\item{...}{Arguments passed to other methods}
}
\value{
An object with new cell names
}
\description{
Change the cell names in all the different parts of an object. Can be useful
before combining multiple objects.
}
\details{
If \code{add.cell.id} is set a prefix is added to existing cell names. If
\code{new.names} is set these will be used to replace existing names.
}
\keyword{internal}
