% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RowMergeSparseMatrices}
\alias{RowMergeSparseMatrices}
\title{Merge Sparse Matrices by Row}
\usage{
RowMergeSparseMatrices(mat1, mat2)
}
\arguments{
\item{mat1}{First matrix}

\item{mat2}{Second matrix or list of matrices}
}
\value{
Returns a sparse matrix
}
\description{
Merge two or more sparse matrices by rowname.
}
\details{
Shared matrix rows (with the same row name) will be merged, and unshared
rows (with different names) will be filled with zeros in the matrix not
containing the row.
}
\concept{utils}
