% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{SaveSeuratBP}
\alias{SaveSeuratBP}
\title{Save \code{Seurat} Objects as RDS files with BPCells directories in same folder}
\usage{
SaveSeuratBP(
  object,
  filename = NULL,
  destdir = NULL,
  relative = FALSE,
  remove_old = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{\link{Seurat}} object}

\item{destdir}{Destination directory (to include both BPCells directory
and \code{\link{Seurat}} object)}

\item{relative}{Save relative paths instead of absolute ones. This is recommended
if sharing the object folder}

\item{remove_old}{Delete current BPCells directories after moving}

\item{...}{
  Arguments passed on to \code{\link[base:readRDS]{base::saveRDS}}
  \describe{
    \item{\code{ascii}}{a logical.  If \code{TRUE} or \code{NA}, an ASCII
    representation is written; otherwise (default), a binary one is used.
    See the comments in the help for \code{\link[base]{save}}.}
    \item{\code{version}}{the workspace format version to use.  \code{NULL}
    specifies the current default version (3). The only other supported
    value is 2, the default from \R 1.4.0 to \R 3.5.0.}
    \item{\code{compress}}{a logical specifying whether saving to a named file is
    to use \code{"gzip"} compression, or one of \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} to indicate the type of compression to
    be used.  Ignored if \code{file} is a connection.}
    \item{\code{refhook}}{a hook function for handling reference objects.}
  }}

\item{file}{file name for \code{\link{Seurat}} object. defaults to
\code{(paste0(Project(object), ".Rds"))}}
}
\value{
Invisibly returns \code{file}
}
\description{
Save \code{Seurat} Objects as RDS files with BPCells directories in same folder
}
\note{
This function requires the
\href{https://cran.r-project.org/package=fs}{\pkg{fs}} package
to be installed
}
\section{Progress Updates with \pkg{progressr}}{

This function uses
\href{https://cran.r-project.org/package=progressr}{\pkg{progressr}} to
render status updates and progress bars. To enable progress updates, wrap
the function call in \code{\link[progressr]{with_progress}} or run
\code{\link[progressr:handlers]{handlers(global = TRUE)}} before running
this function. For more details about \pkg{progressr}, please read
\href{https://progressr.futureverse.org/articles/progressr-intro.html}{\code{vignette("progressr-intro")}}
}

\examples{
if (requireNamespace("BPCells") && requireNamespace("fs")) {
  pbmc_small[["RNA5"]] <- as(object = pbmc_small[['RNA']], Class = 'Assay5')
  pbmc_small[["RNA5"]]$counts <- BPCells::write_matrix_dir(
  mat = pbmc_small[['RNA']]$counts,
    dir = '~/pbmc_counts_BP/')

  # Save `pbmc_small` to a folder with Rds file and BP Cells directory
  SaveSeuratBP(pbmc_small,
               filename = "pbmc_small.Rds",
               destdir = "~/full_object/",
               relative = TRUE,
               remove_old = FALSE)

  # Load the saved object with on-disk layers back into memory
  pbmc2 <- readRDS("~/full_object/pbmc_small.Rds")
  pbmc2
  pbmc2[["RNA5"]]$counts
}

}
\seealso{
\code{\link{saveRDS}()} \code{\link{readRDS}()}
}
