% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{SaveSeuratRDS}
\alias{SaveSeuratRDS}
\title{Save Seurat Objects to RDS files}
\usage{
SaveSeuratRDS(
  object,
  file = NULL,
  absolute = TRUE,
  mv.tmp = TRUE,
  wd = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{\link{Seurat}} object}

\item{file}{Path to save \code{object} to; defaults to
\code{file.path(getwd(), paste0(Project(object), ".Rds"))}}

\item{absolute}{For on-disk layers, store absolute paths instead of
relative ones}

\item{mv.tmp}{For on-disk layers saved in \dQuote{\code{\Sexpr[stage=render]{tempdir()}}}}
}
\description{
Save Seurat Objects to RDS files
}
