% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ScaleFactors}
\alias{ScaleFactors}
\title{Get image scale factors}
\usage{
ScaleFactors(object, ...)
}
\arguments{
\item{object}{An object to get scale factors from}

\item{...}{Arguments passed to other methods}
}
\value{
An object of class \code{scalefactors}
}
\description{
Get image scale factors
}
\concept{spatial}
