% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\docType{class}
\name{Segmentation-class}
\alias{Segmentation-class}
\title{The \code{Segmentation} Class}
\description{
The \code{Segmentation} Class
}
\seealso{
\code{Segmentation} methods: \code{\link{Segmentation-methods}}

Segmentation layer classes:
\code{\link{Centroids-class}},
\code{\link{Molecules-class}}
}
\concept{segmentation}
