% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{Segmentation-methods}
\alias{Segmentation-methods}
\alias{Cells.Segmentation}
\alias{GetTissueCoordinates.Segmentation}
\alias{lengths.Segmentation}
\alias{subset.Segmentation}
\alias{[,Segmentation,ANY,ANY,ANY-method}
\alias{coordinates,Segmentation-method}
\alias{show,Segmentation-method}
\title{\code{Segmentation} Methods}
\usage{
\method{Cells}{Segmentation}(x, ...)

\method{GetTissueCoordinates}{Segmentation}(object, full = TRUE, ...)

\method{lengths}{Segmentation}(x, use.names = TRUE)

\method{subset}{Segmentation}(x, cells = NULL, ...)

\S4method{[}{Segmentation,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{coordinates}{Segmentation}(obj, full = TRUE, ...)

\S4method{show}{Segmentation}(object)
}
\arguments{
\item{x, object, obj}{A
\code{\link[SeuratObject:Segmentation-class]{Segmentation}} object}

\item{...}{Arguments passed to other methods}

\item{full}{Expand the coordinates to the full polygon}

\item{use.names}{Ignored}

\item{cells}{A vector of cells to keep; if \code{NULL}, defaults
to all cells}

\item{i}{A vector of cells to keep; if \code{NULL}, defaults
to all cells}

\item{j}{Ignored}

\item{drop}{Ignored}
}
\value{
\code{Cells}: A vector of cell names

\code{GetTissueCoordinates}, \code{coordinates}: A data frame with
three columns:
\itemize{
 \item \dQuote{\code{x}}: the x-coordinate
 \item \dQuote{\code{y}}: the y-coordinate
 \item \dQuote{\code{cell}} or \dQuote{\code{ID}}: the cell name
}
If \code{full} is \code{TRUE}, then each coordinate will indicate a vertex
for the cell polygon; otherwise, each coordinate will indicate a centroid
for the cell. Note: \code{GetTissueCoordinates} ....

\code{lengths}: An \code{\link[base:rle]{rle}} object for the cells

\code{subset}, \code{[}: \code{x} subsetted to the cells specified
by \code{cells}/\code{i}

\code{show}: Invisibly returns \code{NULL}
}
\description{
Methods for \code{\link[SeuratObject:Segmentation-class]{Segmentation}}
objects
}
\details{
\code{Cells}: Get cell names

\code{GetTissueCoordinates}, \code{coordinates}: Get
tissue coordinates

\code{lengths}: Generate a run-length encoding of the cells present

\code{subset}, \code{[}: Subset a \code{Segmentation} object to
certain cells

\code{show}: Display an object summary to stdout
}
\seealso{
\code{\link{Segmentation-class}}
}
