% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{Segmentation-validity}
\alias{Segmentation-validity}
\title{Segmentation Validity}
\description{
Validation of \code{Segmentation} objects is handled by
\code{\link[methods]{validObject}}
}
\section{sf.data Validation}{

Validates that the sf.data slot contains an object of class \code{sf}.
}

\seealso{
\code{\link[methods]{validObject}}

Segmentation layer classes:
\code{\link{Centroids-class}},
\code{\link{Centroids-methods}},
\code{\link{Molecules-class}},
\code{\link{Molecules-methods}},
\code{\link{Segmentation-class}},
\code{\link{Segmentation-methods}}
}
\concept{segmentation}
