% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatialcoords.R
\name{Segmentations}
\alias{Segmentations}
\alias{DefaultSegmentation}
\alias{DefaultSegmentation<-}
\alias{Molecules}
\alias{DefaultSegmentation.SpatialCoords}
\alias{DefaultSegmentation<-.SpatialCoords}
\alias{Molecules.SpatialCoords}
\alias{Segmentations.SpatialCoords}
\title{Get, Set, and Query Segmentation Layers}
\usage{
DefaultSegmentation(object)

DefaultSegmentation(object, ...) <- value

Molecules(object, ...)

Segmentations(object, ...)

\method{DefaultSegmentation}{SpatialCoords}(object)

\method{DefaultSegmentation}{SpatialCoords}(object, ...) <- value

\method{Molecules}{SpatialCoords}(object, ...)

\method{Segmentations}{SpatialCoords}(object, molecules = FALSE, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{The name of a segmentation layer to set as default}

\item{molecules}{Include \dQuote{molecules} as a segmentation layer, if
molecule information is present}
}
\value{
\code{DefaultSegmentation}: The name of the default
segmentation layer

\code{DefaultSegmentations<-}: \code{object} with the default
segmentation layer set to \code{value}

\code{Molecules}: The names of all molecule sets present within
\code{object}

\code{Segmentations}: The names of all segmentation layers present
within \code{object}
}
\description{
Get, Set, and Query Segmentation Layers
}
