% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R
\name{SetAssayData}
\alias{SetAssayData}
\alias{SetAssayData.Assay}
\title{Setter for multimodal data}
\usage{
SetAssayData(object, ...)

\method{SetAssayData}{Assay}(object, slot = c("counts", "data", "scale.data"), new.data, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{slot}{Where to store the new data}

\item{new.data}{New data to insert}
}
\value{
object with the assay data set
}
\description{
Setter for multimodal data
}
\examples{
# Set an Assay slot directly
count.data <- GetAssayData(object = pbmc_small[["RNA"]], slot = "counts")
count.data <- as.matrix(x = count.data + 1)
new.assay <- SetAssayData(object = pbmc_small[["RNA"]], slot = "counts", new.data = count.data)

}
\concept{data-access}
