% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\docType{class}
\name{Seurat-class}
\alias{Seurat-class}
\alias{Seurat}
\title{The Seurat Class}
\description{
The Seurat object is a representation of single-cell expression data for R;
each Seurat object revolves around a set of cells and consists of one or more
\code{\link{Assay}} objects, or individual representations of
expression data (eg. RNA-seq, ATAC-seq, etc). These assays can be reduced
from their high-dimensional state to a lower-dimension state and stored as
\code{\link{DimReduc}} objects. Seurat objects also
store additional metadata, both at the cell and feature level (contained
within individual assays). The object was designed to be as self-contained as
 possible, and easily extendable to new methods.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of assays for this project}

\item{\code{meta.data}}{Contains meta-information about each cell, starting with
number of features detected (nFeature) and the original identity class
(orig.ident); more information is added using \code{\link{AddMetaData}}}

\item{\code{active.assay}}{Name of the active, or default, assay; settable using
\code{\link{DefaultAssay}}}

\item{\code{active.ident}}{The active cluster identity for this Seurat object;
settable using \code{\link{Idents}}}

\item{\code{graphs}}{A list of \code{\link{Graph}} objects}

\item{\code{neighbors}}{...}

\item{\code{reductions}}{A list of dimensional reduction objects for this object}

\item{\code{images}}{A list of spatial image objects}

\item{\code{project.name}}{Name of the project}

\item{\code{misc}}{A list of miscellaneous information}

\item{\code{version}}{Version of Seurat this object was built under}

\item{\code{commands}}{A list of logged commands run on this \code{Seurat} object}

\item{\code{tools}}{A list of miscellaneous data generated by other tools, should be
filled by developers only using \code{\link{Tool}<-}}
}}

\seealso{
Other seurat: 
\code{\link{$.Seurat}()},
\code{\link{Seurat-validity}},
\code{\link{[[.Seurat}()},
\code{\link{[[<-,Seurat}},
\code{\link{names.Seurat}()},
\code{\link{subset.Seurat}()}
}
\concept{seurat}
