% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{Seurat-methods}
\alias{Seurat-methods}
\alias{.DollarNames.Seurat}
\alias{$.Seurat}
\alias{$<-.Seurat}
\alias{[.Seurat}
\alias{[[.Seurat}
\alias{dim.Seurat}
\alias{dimnames.Seurat}
\alias{names.Seurat}
\alias{[[<-,Seurat-method}
\alias{show,Seurat-method}
\title{Seurat Methods}
\usage{
\method{.DollarNames}{Seurat}(x, pattern = "")

\method{$}{Seurat}(x, i, ...)

\method{$}{Seurat}(x, i, ...) <- value

\method{[}{Seurat}(x, i, j, ...)

\method{[[}{Seurat}(x, i, ..., drop = FALSE)

\method{dim}{Seurat}(x)

\method{dimnames}{Seurat}(x)

\method{names}{Seurat}(x)

\S4method{[[}{Seurat}(x, i, j, ...) <- value

\S4method{show}{Seurat}(object)
}
\arguments{
\item{x, object}{A \code{\link{Seurat}} object}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }

\item{i, features}{Depends on the method
\describe{
 \item{\code{[}, \code{subset}}{Feature names or indices}
 \item{\code{$}, \code{$<-}}{Name of a single metadata column}
 \item{\code{[[}, \code{[[<-}}{
  Name of one or more metadata columns or an associated object; associated
  objects include \code{\link{Assay}}, \code{\link{DimReduc}},
  \code{\link{Graph}}, \code{\link{SeuratCommand}}, or
  \code{\link{SpatialImage}} objects
 }
 @param j,cells Cell names or indices
 @param ... Arguments passed to other methods
}}

\item{value}{Additional metadata or associated objects to add; \strong{note}:
can pass \code{NULL} to remove metadata or an associated object}

\item{drop}{See \code{\link[base]{drop}}}
}
\value{
\code{$}: metadata column \code{i} for object \code{x};
\strong{note}: unlike \code{[[}, \code{$} drops the shape of the metadata
to return a vector instead of a dataframe

\code{$<-}: object \code{x} with metadata \code{value} saved as
\code{i}

\code{[}: object \code{x} with features \code{i} and cells \code{j}

\code{[[}: If \code{i} is missing, the metadata dataframe; if
\code{i} is a vector of metadata names, a dataframe with the requested
metadata, otherwise, the requested associated object

\code{dim}: The number of features (\code{nrow}) and cells
(\code{ncol}) for the default assay; \strong{note}: while the number of
features changes depending on the active assay, the number of cells remains
the same across all assays

\code{dimnames}: The feature (row) and cell (column) names;
\strong{note}: while the features change depending on the active assay, the
cell names remain the same across all assays

\code{names}: The names of all \code{\link{Assay}},
\code{\link{DimReduc}}, \code{\link{Graph}}, and \code{\link{SpatialImage}}
objects in the \code{Seurat} object

\code{[[<-}: \code{x} with the metadata or associated objects added
as \code{i}; if \code{value} is \code{NULL}, removes metadata or associated
object \code{i} from object \code{x}

\code{show}: Prints summary to \code{\link[base]{stdout}} and
invisibly returns \code{NULL}
}
\description{
Methods for \code{\link{Seurat}} objects for generics defined in other
packages
}
\section{Functions}{
\itemize{
\item \code{.DollarNames.Seurat}: Autocompletion for \code{$} access on a
\code{Seurat} object

\item \code{$.Seurat}: Metadata access for \code{Seurat} objects

\item \code{$<-.Seurat}: Metadata setter for \code{Seurat} objects

\item \code{[.Seurat}: Simple subsetter for \code{Seurat} objects

\item \code{[[.Seurat}: Metadata and associated object accessor

\item \code{dim.Seurat}: Number of cells and features for the active assay

\item \code{dimnames.Seurat}: The cell and feature names for the active assay

\item \code{names.Seurat}: Common associated objects

\item \code{[[<-,Seurat-method}: Add cell-level metadata or associated objects

\item \code{show,Seurat-method}: Overview of a \code{Seurat} object
}}

\examples{
pbmc_small[VariableFeatures(object = pbmc_small), ]
pbmc_small[, 1:10]

}
\concept{seurat}
