% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{Seurat-validity}
\alias{Seurat-validity}
\title{Seurat Object Validity}
\description{
Validation of \code{Seurat} objects is handled by
\code{\link[methods]{validObject}}
}
\seealso{
\code{\link[methods]{validObject}}

Seurat object, validity, and interaction methods
\code{\link{$.Seurat}()},
\code{\link{Seurat-class}},
\code{\link{[[.Seurat}()},
\code{\link{[[<-,Seurat,NULL}},
\code{\link{[[<-,Seurat}},
\code{\link{dim.Seurat}()},
\code{\link{dimnames.Seurat}()},
\code{\link{merge.Seurat}()},
\code{\link{names.Seurat}()},
\code{\link{subset.Seurat}()}
}
\concept{seurat}
