% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{SeuratCommand-methods}
\alias{SeuratCommand-methods}
\alias{.DollarNames.SeuratCommand}
\alias{$.SeuratCommand}
\alias{[.SeuratCommand}
\alias{as.list.SeuratCommand}
\alias{show,SeuratCommand-method}
\title{\code{SeuratCommand} Methods}
\usage{
\method{.DollarNames}{SeuratCommand}(x, pattern = "")

\method{$}{SeuratCommand}(x, i, ...)

\method{[}{SeuratCommand}(x, i, ...)

\method{as.list}{SeuratCommand}(x, complete = FALSE, ...)

\S4method{show}{SeuratCommand}(object)
}
\arguments{
\item{x, object}{A \code{\link{SeuratCommand}} object}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }

\item{i}{For a \code{$}, a parameter name; for \code{[}, a
\code{SeuratCommand} slot name}

\item{...}{Arguments passed to other methods}

\item{complete}{Include slots besides just parameters
(eg. call string, name, timestamp)}
}
\value{
\code{$}: The value for parameter \code{i}

\code{[}: Slot \code{i} from \code{x}

\code{as.list}: A list with the parameters and, if
\code{complete = TRUE}, the call string, name, and timestamp

\code{show}: Prints summary to \code{\link[base]{stdout}} and
invisibly returns \code{NULL}
}
\description{
Methods for \code{\link{SeuratCommand}} objects for generics defined in
other packages
}
\section{Functions}{
\itemize{
\item \code{.DollarNames(SeuratCommand)}: Autocompletion for \code{$} access on a
\code{SeuratCommand} object

\item \code{$}: Access a parameter from a
\code{SeuratCommand} object

\item \code{[}: Access data from a \code{SeuratCommand}
object

\item \code{as.list(SeuratCommand)}: Coerce a \code{SeuratCommand} to a list

\item \code{show(SeuratCommand)}: Overview of a \code{SeuratCommand} object

}}
\concept{command}
