% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{SeuratObject-options}
\alias{SeuratObject-options}
\title{\pkg{Seurat} Options}
\description{
Various options used in \pkg{Seurat}
}
\section{Package Options}{

\subsection{Seurat.coords.short_range}{
Defaults to
\dQuote{\Sexpr[stage=build]{SeuratObject:::Seurat.options$Seurat.coords.short_range}}\cr
Currently set to \dQuote{\Sexpr[stage=render]{getOption("Seurat.coords.short_range")}}
}
\subsection{Seurat.input.sparse_ratio}{
Defaults to
\dQuote{\Sexpr[stage=build]{SeuratObject:::Seurat.options$Seurat.input.sparse_ratio}}\cr
Currently set to \dQuote{\Sexpr[stage=render]{getOption("Seurat.input.sparse_ratio")}}
}
\subsection{Seurat.io.rds.strict}{
Defaults to
\dQuote{\Sexpr[stage=build]{SeuratObject:::Seurat.option$Seurat.io.rds.strict}}\cr
Currently set to \dQuote{\Sexpr[stage=render](getOption("Seurat.io.rds.strict"))}
}
\subsection{Seurat.object.assay.calcn}{
Run \code{CalcN} when adding assay data to a \code{Seurat} object\cr
Defaults to
\dQuote{\Sexpr[stage=build]{SeuratObject:::Seurat.options$Seurat.object.assay.calcn}}\cr
Currently set to \dQuote{\Sexpr[stage=render]{getOption("Seurat.object.assay.calcn")}}
}
\subsection{Seurat.object.assay.version}{
Defaults to
\dQuote{\Sexpr[stage=build]{SeuratObject:::Seurat.options$Seurat.object.assay.version}}\cr
Currently set to \dQuote{\Sexpr[stage=render]{getOption("Seurat.object.assay.version")}}
}
\subsection{Seurat.object.assay.v3.missing_layer}{
Defaults to
\dQuote{\Sexpr[stage=build]{SeuratObject:::Seurat.options$Seurat.object.assay.v3.missing_layer}}\cr
Currently set to \dQuote{\Sexpr[stage=render]{getOption("Seurat.object.assay.v3.missing_layer")}}
}
\subsection{Seurat.object.project}{
Default project for new \code{\link{Seurat}} objects\cr
Defaults to
\dQuote{\Sexpr[stage=build]{SeuratObject:::Seurat.options$Seurat.object.project}}\cr
Currently set to \dQuote{\Sexpr[stage=render]{getOption("Seurat.object.project")}}
}
}

\keyword{internal}
