% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratObject-package}
\alias{SeuratObject}
\alias{SeuratObject-package}
\title{SeuratObject: Data Structures for Single Cell Data}
\description{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.satijalab.org/seurat}
  \item \url{https://github.com/satijalab/seurat-object}
  \item Report bugs at \url{https://github.com/satijalab/seurat-object/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{nygcSatijalab@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Authors:
\itemize{
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID})
  \item Andrew Butler \email{abutler@nygenome.org} (\href{https://orcid.org/0000-0003-3608-0463}{ORCID})
  \item Tim Stuart \email{tstuart@nygenome.org} (\href{https://orcid.org/0000-0002-3044-0897}{ORCID})
}

Other contributors:
\itemize{
  \item Jeff Farrell \email{jfarrell@g.harvard.edu} [contributor]
  \item Shiwei Zheng \email{szheng@nygenome.org} (\href{https://orcid.org/0000-0001-6682-6743}{ORCID}) [contributor]
  \item Christoph Hafemeister \email{chafemeister@nygenome.org} (\href{https://orcid.org/0000-0001-6365-8254}{ORCID}) [contributor]
  \item Patrick Roelli \email{proelli@nygenome.org} [contributor]
  \item Yuhan Hao \email{yhao@nygenome.org} (\href{https://orcid.org/0000-0002-1810-0822}{ORCID}) [contributor]
}

}
