% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{show,Assay-method}
\alias{show,Assay-method}
\title{V3 Assay Overview}
\usage{
\S4method{show}{Assay}(object)
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly
returns \code{NULL}
}
\description{
Overview of an \code{\link{Assay}} object
}
\examples{
rna <- pbmc_small[["RNA"]]
rna

}
\seealso{
\code{\link{Assay}}
}
\concept{assay}
\keyword{internal}
