% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R
\name{show,DimReduc-method}
\alias{show,DimReduc-method}
\title{Dimensional Reduction Overview}
\usage{
\S4method{show}{DimReduc}(object)
}
\arguments{
\item{object}{A dimensional reduction}
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly
returns \code{NULL}
}
\description{
Overview of a \code{\link{DimReduc}} object
}
\examples{
pca <- pbmc_small[["pca"]]
pca

}
\seealso{
\code{\link{DimReduc}}
}
\keyword{internal}
