% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{show,Graph-method}
\alias{show,Graph-method}
\title{Graph Object Overview}
\usage{
\S4method{show}{Graph}(object)
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly
returns \code{NULL}
}
\description{
Overview of a \code{\link{Graph}} Object
}
\examples{
pbmc_small[["RNA_snn"]]

}
\concept{graph}
\keyword{internal}
