% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmap.R
\name{show,LogMap-method}
\alias{show,LogMap-method}
\title{\code{\link{LogMap}} Object Overview}
\usage{
\S4method{show}{LogMap}(object)
}
\arguments{
\item{object}{A \code{\link{LogMap}} object}
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly
returns \code{NULL}
}
\description{
Overview of a \code{\link{LogMap}} object
}
\concept{logmap}
