% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{show,seurat-method}
\alias{show,seurat-method}
\title{Old Seurat Object Overview}
\usage{
\S4method{show}{seurat}(object)
}
\arguments{
\item{object}{An old seurat object}
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly
returns \code{NULL}
}
\description{
Overview of a \code{\link[=oldseurat]{seurat}} object overview
}
\concept{oldseurat}
\keyword{internal}
