% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{show,Seurat-method}
\alias{show,Seurat-method}
\title{Seurat Object Overview}
\usage{
\S4method{show}{Seurat}(object)
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly
returns \code{NULL}
}
\description{
Overview of a \code{\link{Seurat}} object
}
\examples{
pbmc_small

}
\concept{seurat}
\keyword{internal}
