% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{show,SeuratCommand-method}
\alias{show,SeuratCommand-method}
\title{Command Log Overview}
\usage{
\S4method{show}{SeuratCommand}(object)
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly
returns \code{NULL}
}
\description{
Overview of a \code{\link{SeuratCommand}} object
}
\examples{
cmd <- pbmc_small[["NormalizeData.RNA"]]
cmd

}
\concept{command}
\keyword{internal}
