% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{show,StdAssay-method}
\alias{show,StdAssay-method}
\title{V5 Assay Overview}
\usage{
\S4method{show}{StdAssay}(object)
}
\arguments{
\item{object}{A v5 Assay}
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly
returns \code{NULL}
}
\description{
Overview of a \code{\link{StdAssay}} object
}
\seealso{
Other stdassay: 
\code{\link{StdAssay-class}},
\code{\link{StdAssay-validity}},
\code{\link{[.StdAssay}()},
\code{\link{[[.StdAssay}()},
\code{\link{dim.StdAssay}()},
\code{\link{dimnames.StdAssay}()},
\code{\link{subset.StdAssay}()}
}
\concept{stdassay}
\keyword{internal}
