% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Simplify}
\alias{Simplify}
\title{Simplify Geometry}
\usage{
Simplify(coords, tol, topologyPreserve = TRUE)
}
\arguments{
\item{coords}{...}

\item{tol}{Numerical tolerance value to be used by the Douglas-Peuker algorithm}

\item{topologyPreserve}{Logical determining if the algorithm should attempt to preserve the topology of the original geometry}
}
\value{
...
}
\description{
Simplify Geometry
}
