% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SparseEmptyMatrix}
\alias{SparseEmptyMatrix}
\title{Generate empty dgC sparse matrix}
\usage{
SparseEmptyMatrix(nrow, ncol, rownames = NULL, colnames = NULL)
}
\arguments{
\item{ncol, nrow}{Number of columns and rows in matrix}

\item{rownames, colnames}{Optional row- and column names for the matrix}
}
\description{
Generate empty dgC sparse matrix
}
\keyword{internal}
