% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcoords.R
\docType{class}
\name{SpatialCoords-class}
\alias{SpatialCoords-class}
\alias{SpatialCoords}
\title{The Spirula Coordinates Object}
\description{
A modern container for storing coordinates of spatially-resolved single
cells. Capable of storing multiple layers of centroid/voxel- and
polygon/segmentation-based coordinate systems. Supports coordinates for
spatially-resolved molecule (FISH) data. Compatible with Seurat's
\code{\link{SpatialImage}}
}
\section{Slots}{

\describe{
\item{\code{cells}}{(\code{\link{LogMap}}) A map of the cells present
in each segmentation layer}

\item{\code{molecules}}{(\code{\link[base]{list}}) A named list of
\code{\link{Molecules}} objects defining spatially-resolved
molecular coordinates}

\item{\code{segmentations}}{(\code{[named]\link[base]{list}}
\{\code{\link{Segmentation}}, \code{\link{Centroids}}\}) A named list of
\code{\link{Segmentation}} and \code{\link{Centroids}} objects defining
spatially-resolved cellular segmentations or centroids}

\item{\code{assay}}{(\code{\link[base:character]{character [1L]}}) A character naming
the associated assay of the spatial coordinates}

\item{\code{key}}{(\code{\link[base:character]{character [1L]}}) The key for the
spatial coordinates}
}}

\seealso{
\code{\link{SpatialCoords-methods}}
}
