% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcoords.R
\name{SpatialCoords-methods}
\alias{SpatialCoords-methods}
\alias{Cells.SpatialCoords}
\alias{Features.SpatialCoords}
\alias{FetchData.SpatialCoords}
\alias{GetTissueCoordinates.SpatialCoords}
\alias{Keys.SpatialCoords}
\alias{$.SpatialCoords}
\alias{[.SpatialCoords}
\alias{[[.SpatialCoords}
\alias{length.SpatialCoords}
\alias{names.SpatialCoords}
\alias{[[<-,SpatialCoords,character,missing,Centroids-method}
\alias{[[<-,SpatialCoords,character,missing,Molecules-method}
\alias{[[<-,SpatialCoords,character,missing,NULL-method}
\alias{[[<-,SpatialCoords,character,missing,Segmentation-method}
\alias{show,SpatialCoords-method}
\title{\code{SpatialCoords} Methods}
\usage{
\method{Cells}{SpatialCoords}(x, layer = NULL, ...)

\method{Features}{SpatialCoords}(x, layer = NULL, ...)

\method{FetchData}{SpatialCoords}(object, vars, cells = NULL, simplify = TRUE, ...)

\method{GetTissueCoordinates}{SpatialCoords}(object, which = NULL, ...)

\method{Keys}{SpatialCoords}(object, ...)

\method{$}{SpatialCoords}(x, i, ...)

\method{[}{SpatialCoords}(x, i, j, ...)

\method{[[}{SpatialCoords}(x, i, ...)

\method{length}{SpatialCoords}(x)

\method{names}{SpatialCoords}(x)

\S4method{[[}{SpatialCoords,character,missing,Centroids}(x, i, j, ...) <- value

\S4method{[[}{SpatialCoords,character,missing,Molecules}(x, i, j, ...) <- value

\S4method{[[}{SpatialCoords,character,missing,`NULL`}(x, i, j, ...) <- value

\S4method{[[}{SpatialCoords,character,missing,Segmentation}(x, i, j, ...) <- value

\S4method{show}{SpatialCoords}(object)
}
\arguments{
\item{x, object}{A \code{\link{SpatialCoords}} object}

\item{layer}{Name of segmentation or molecular layer to extract cell or
feature names for; pass \code{NA} to return all cells or feature names}

\item{...}{Arguments passed to other methods}

\item{vars}{A vector of variables to fetch; can be the name of a
segmentation layer, to get tissue coordinates, or molecule names,
to get molecule coordinates}

\item{simplify}{If only returning either tissue or molecule coordinates,
return a single data frame instead of a list}

\item{i, cells}{For \code{[[} and \code{[[<-}, the name of a segmentation or
\dQuote{molecules}; for \code{FetchData}, \code{subset}. and \code{[}, a
vector of cells to keep}

\item{j, features}{For \code{subset} and \code{[}, a vector of features to
keep; for \code{[[<-}, not used}

\item{value}{For \code{[[<-}, a replacement \code{\link{Molecules}},
\code{\link{Centroids}}, or \code{\link{Segmentations}} object or \code{NULL}
to remove the layer stored at \code{i}}
}
\value{
\code{Cells}: A vector of cell names

\code{Features}: A vector of spatially-resolved molecule names;
if no molecular information present, returns \code{NULL}

\code{FetchData}: If both molecule and tissue coordinates are
requested, then a two-length list:
\itemize{
 \item \dQuote{\code{molecules}}: A data frame with the molecule coordinates
  requested. If molecules requested are keyed, the keys are preserved in the
  data frame
 \item \dQuote{\code{coordinates}}: A data frame with coordinates from the
  segmentation layers requested
}
If \code{simplify} is \code{TRUE} and only one data frame is generated, then
only the data frame is returned. Otherwise, a one-length list is returned
with the single data frame generated

\code{GetTissueCoordinates}: ...

\code{Keys}: A named vector of molecule set keys; names are the
names of the molecule sets and values are the keys for the respective
molecule set

\code{$}, \code{[[}: The segmentation layer or spatially-resolved
molecule information stored at \code{i}

\code{length}: The number of segmentation layers
(\code{\link{Segmentation}} or \code{\link{Centroids}} objects)

\code{names}: A vector of segmentation layer names

\code{[[<-}: Varies depending on the class of \code{value}:
\itemize{
 \item If \code{value} is \code{NULL}, returns \code{x} with the layer
 \code{i} removed; also allows removing \code{molecules}; does not allow
 removing the default segmentation
 \item If \code{value} is a \code{Molecules}, returns \code{x} with
 \code{value} stored in \code{molecules}; requires that \code{i} is
 \dQuote{molecules}
 \item Otherwise, stores \code{value} as a segmentation layer named \code{i}
}

\code{show}: Invisibly returns \code{NULL}
}
\description{
Methods for \code{\link{SpatialCoords}} objects
}
\details{
The following methods are defined for interacting with a
\code{SpatialCoords} object:

\code{Cells}: Get cell names

\code{Features}: Get spatially-resolved molecule names

\code{FetchData}: Fetch tissue and/or molecule coordinates from
a \code{SpatialCoords} object

\code{GetTissueCoordinates}: Get cell or molecule  coordinates from
a \code{SpatialCoords} object

\code{Keys}: Get the keys of molecule sets contained within a
\code{SpatialCoords} object

\code{$}, \code{[[}: Extract a segmentation layer

\code{length}: Get the number of segmentation layers in a
\code{SpatialCoords} object

\code{names}: Get the names of segmentation layers, equivalent to
\code{\link[Spirula:Segmentations]{Segmentations(x, molecules = FALSE)}}

\code{[[<-}: Add or remove segmentation layers and molecule
information to/from a \code{SpatialCoords} object

\code{show}: Display an object summary to stdout
}
\seealso{
\code{\link{SpatialCoords-class}}
}
