% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\docType{class}
\name{SpatialImage-class}
\alias{SpatialImage-class}
\alias{SpatialImage}
\title{The SpatialImage class}
\description{
The \code{SpatialImage} class is a virtual class representing spatial
information for Seurat. All spatial image information must inherit from this
class for use with \code{Seurat} objects
}
\section{Slots}{

\describe{
\item{\code{assay}}{Name of assay to associate image data with; will give this image
priority for visualization when the assay is set as the active/default assay
in a \code{Seurat} object}

\item{\code{key}}{A one-length character vector with the object's key; keys must
be one or more alphanumeric characters followed by an underscore
\dQuote{\code{_}} (regex pattern
\dQuote{\code{\Sexpr[stage=build]{SeuratObject:::.KeyPattern()}}})}

\item{\code{misc}}{A named list of unstructured miscellaneous data}
}}

\seealso{
\code{\link{SpatialImage-methods}} for a list of required and
provided methods
}
