% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R, R/seurat.R
\name{SpatiallyVariableFeatures}
\alias{SpatiallyVariableFeatures}
\alias{SpatiallyVariableFeatures.Assay}
\alias{SpatiallyVariableFeatures.Seurat}
\title{Get spatially variable feature information}
\usage{
SpatiallyVariableFeatures(object, ...)

\method{SpatiallyVariableFeatures}{Assay}(
  object,
  selection.method = "markvariogram",
  decreasing = TRUE,
  ...
)

\method{SpatiallyVariableFeatures}{Seurat}(
  object,
  assay = NULL,
  selection.method = "markvariogram",
  decreasing = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{selection.method}{Method used to set variable features}

\item{decreasing}{Return features in decreasing order (most spatially
variable first).}

\item{assay}{Name of assay to pull variable features for}
}
\description{
Get spatially variable feature information
}
\concept{data-access}
