% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay5.R, R/seurat.R
\name{JoinLayers}
\alias{JoinLayers}
\alias{JoinLayers.Assay5}
\alias{JoinLayers.Seurat}
\title{Split and Join Layers Together}
\usage{
JoinLayers(object, ...)

\method{JoinLayers}{Assay5}(object, layers = NULL, new = NULL, ...)

\method{JoinLayers}{Seurat}(object, assay = NULL, layers = NULL, new = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{layers}{Names of layers to split or join}

\item{new}{Name of new layers}
}
\value{
\code{object} with the layers specified joined
}
\description{
Split and Join Layers Together
}
