% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\docType{class}
\name{StdAssay-class}
\alias{StdAssay-class}
\alias{StdAssay}
\title{Core Assay Infrastructure}
\description{
The \code{StdAssay} class is a virtual class that provides core
infrastructure for assay data in \pkg{Seurat}. Assays contain expression
data (layers) and associated feature-level metadata. Derived classes
(eg. \link[Assay-class]{the v5 Assay}) may optionally define additional
functionality
}
\section{Slots}{

\describe{
\item{\code{layers}}{A named list containing alternate representations of
\code{data}; layers must have the same cells as \code{data} and either the
same or a subset of the features present in \code{data}}

\item{\code{cells}}{A \code{\link{LogMap}} describing the cell membership for
each layer}

\item{\code{features}}{A \code{\link{LogMap}} describing the feature membership
for each layer}

\item{\code{assay.orig}}{...}

\item{\code{meta.data}}{...}

\item{\code{misc}}{...}

\item{\code{key}}{A one-length character vector with the object's key; keys must
be one or more alphanumeric characters followed by an underscore
\dQuote{\code{_}} (regex pattern \dQuote{\code{^[[:alnum:]]+_$}})}
}}

\seealso{
\code{\link{Assay5-class}}
}
