% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{StdAssay-methods}
\alias{StdAssay-methods}
\alias{[.StdAssay}
\alias{[[.StdAssay}
\alias{dim.StdAssay}
\alias{dimnames.StdAssay}
\alias{head.StdAssay}
\alias{tail.StdAssay}
\alias{[[<-,StdAssay,ANY,ANY,ANY-method}
\alias{show,StdAssay-method}
\title{\code{StdAssay} Methods}
\usage{
\method{[}{StdAssay}(x, i, j, ...)

\method{[[}{StdAssay}(x, i, ..., drop = FALSE)

\method{dim}{StdAssay}(x)

\method{dimnames}{StdAssay}(x)

\method{head}{StdAssay}(x, n = 10L, ...)

\method{tail}{StdAssay}(x, n = 10L, ...)

\S4method{[[}{StdAssay,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{show}{StdAssay}(object)
}
\arguments{
\item{x, object}{An \code{\link{StdAssay}} object}

\item{i, features}{For \code{[[}: metadata names; for all other methods,
feature names or indices}

\item{j, cells}{Cell names or indices}

\item{...}{Arguments passed to other methods}

\item{drop}{See \code{\link[base]{drop}}}

\item{value}{Additional metadata to add}
}
\value{
\code{[}: The \code{data} slot for features \code{i} and cells
\code{j}

\code{[[}: The feature-level metadata for \code{i}

\code{dim}: The number of features (\code{nrow}) and cells
(\code{ncol})

\code{dimnames}: Feature (row) and cell (column) names

\code{head}: The first \code{n} rows of feature-level metadata

\code{tail}: The last \code{n} rows of feature-level metadata

\code{[[<-}: \code{x} with metadata \code{value} added as \code{i}

\code{show}: Prints summary to \code{\link[base]{stdout}} and
invisibly returns \code{NULL}
}
\description{
Methods for \code{\link{StdAssay}} objects for generics defined in
other packages
}
\section{Functions}{
\itemize{
\item \code{[.StdAssay}: Get expression data from an \code{StdAssay}

\item \code{[[.StdAssay}: Get feature-level metadata

\item \code{dim.StdAssay}: Number of cells and features for an \code{StdAssay}

\item \code{dimnames.StdAssay}: Cell- and feature-names for an \code{StdAssay}

\item \code{head.StdAssay}: Get the first rows of feature-level metadata

\item \code{tail.StdAssay}: Get the last rows of feature-level metadata

\item \code{[[<-,StdAssay,ANY,ANY,ANY-method}: Add feature-level metadata

\item \code{show,StdAssay-method}: Overview of an \code{StdAssay} object
}}

\concept{assay}
