% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{StdAssay-validity}
\alias{StdAssay-validity}
\title{V5 Assay Validity}
\description{
Validation of \code{StdAssay} objects is handled by
\code{\link[methods]{validObject}}
}
\section{Layer Validation}{

blah
}

\section{Key Validation}{

Keys must be a one-length character vector; a key must be composed of one
of the following:
\itemize{
 \item An empty string (eg. \dQuote{\code{''}}) where \code{nzchar() == 0}
 \item An string composed of one or more alphanumeric values
 (both lower- and upper-case) that ends with an underscore
 (\dQuote{\code{_}}); the first character must be a letter
}
Keys that are not empty strings are validated with the regex
\dQuote{\code{\Sexpr[stage=build]{SeuratObject:::.KeyPattern()}}}
}

\seealso{
\code{\link[methods]{validObject}}

v5 Standard Assay object, validity, and interaction methods
\code{\link{StdAssay-class}},
\code{\link{[.StdAssay}()},
\code{\link{[[.StdAssay}()},
\code{\link{dim.StdAssay}()},
\code{\link{dimnames.StdAssay}()},
\code{\link{show,StdAssay-method}},
\code{\link{subset.StdAssay}()}
}
\concept{stdassay}
\keyword{internal}
