% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Stdev}
\alias{Stdev}
\title{Get the standard deviations for an object}
\usage{
Stdev(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get the standard deviations for an object
}
