% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{StitchMatrix}
\alias{StitchMatrix}
\title{Stitch Matrices Together}
\usage{
StitchMatrix(x, y, rowmap, colmap, ...)
}
\arguments{
\item{x}{A matrix}

\item{y}{One or more matrices of the same class or coercible to the
same class as \code{x}}

\item{rowmap, colmap}{\code{\link{LogMap}s} describing the row and cell
membership of each matrix; the \code{LogMap} entries are assumed to be in
the order of \code{c(x, y)}}

\item{...}{Arguments passed to other methods}
}
\value{
A single matrix of type \code{class(x)} consisting of all values
in component matrices
}
\description{
Stitch Matrices Together
}
\concept{utils}
