% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{[.SeuratCommand}
\alias{[.SeuratCommand}
\title{Command Log Data Access}
\usage{
\method{[}{SeuratCommand}(x, i, ...)
}
\arguments{
\item{x}{A \code{\link{SeuratCommand}} object}

\item{i}{The name of a command log slot}

\item{...}{Ignored}
}
\value{
\code{[}: Slot \code{i} from \code{x}
}
\description{
Access data from a \code{SeuratCommand} object
}
\examples{
cmd <- pbmc_small[["NormalizeData.RNA"]]
cmd["call.string"]

}
\seealso{
Command log object and interaction methods
\code{\link{$.SeuratCommand}()},
\code{\link{.DollarNames.SeuratCommand}()},
\code{\link{LogSeuratCommand}()},
\code{\link{SeuratCommand-class}},
\code{\link{as.list.SeuratCommand}()}
}
\concept{command}
