% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmap.R
\name{[,LogMap}
\alias{[,LogMap}
\alias{[,LogMap,missing,missing,ANY-method}
\alias{[,LogMap,character,character,ANY-method}
\alias{[,LogMap,character,missing,ANY-method}
\alias{[,LogMap,missing,character,ANY-method}
\alias{[,LogMap,numeric,missing,ANY-method}
\alias{[,LogMap,missing,numeric,ANY-method}
\alias{[,LogMap,numeric,numeric,ANY-method}
\title{Matrix-like Subsetting for \link[=LogMap]{Logical Maps}}
\usage{
\S4method{[}{LogMap,missing,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{LogMap,character,character,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{LogMap,character,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{LogMap,missing,character,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{LogMap,numeric,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{LogMap,missing,numeric,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{LogMap,numeric,numeric,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{LogMap} object}

\item{i, j}{...}

\item{...}{Arguments passed to other methods}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }
}
\description{
Matrix-like Subsetting for \link[=LogMap]{Logical Maps}
}
\keyword{internal}
