% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{[[.Assay5}
\alias{[[.Assay5}
\alias{head.Assay5}
\alias{tail.Assay5}
\title{Feature-Level Meta Data}
\usage{
\method{[[}{Assay5}(x, i, ..., drop = FALSE)

\method{head}{Assay5}(x, n = 10L, ...)

\method{tail}{Assay5}(x, n = 10L, ...)
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{i}{Name of feature-level meta data to fetch or add}

\item{...}{Ignored}

\item{drop}{See \code{\link{drop}}}

\item{n}{Number of meta data rows to show}

\item{j}{Ignored}
}
\value{
\code{[[}: The feature-level meta data for \code{i}

\code{head}: The first \code{n} rows of feature-level meta data

\code{tail}: the last \code{n} rows of feature-level meta data
}
\description{
Get and set feature-level meta data
}
\seealso{
v5 Assay object, validity, and interaction methods:
\code{\link{Assay5-class}},
\code{\link{Assay5-validity}},
\code{\link{[.Assay5}()},
\code{\link{dim.Assay5}()},
\code{\link{dimnames.Assay5}()},
\code{\link{merge.Assay5}()},
\code{\link{split.Assay5}()},
\code{\link{subset.Assay5}()}
}
\concept{assay5}
