% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmap.R
\name{[[,LogMap,integer,missing-method}
\alias{[[,LogMap,integer,missing-method}
\alias{[[,LogMap,numeric,missing-method}
\title{\code{\link{LogMap}} Interaction Methods}
\usage{
\S4method{[[}{LogMap,integer,missing}(x, i, j, ...)

\S4method{[[}{LogMap,numeric,missing}(x, i, j, ...)
}
\arguments{
\item{x}{A \code{LogMap} object}

\item{i}{An integer or numeric vector of length 1}

\item{j}{Not used}

\item{...}{Ignored}
}
\value{
The rownames that are mapped to \code{i}
}
\description{
Additional methods for using \code{[[} with \code{\link{LogMap}} objects
}
\keyword{internal}
