% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{[[<-,Seurat,NULL}
\alias{[[<-,Seurat,NULL}
\alias{remove-object}
\alias{remove-objects}
\alias{\S4method{[[<-}{Seurat,character,missing,NULL}}
\alias{[[<-,Seurat,character,missing,NULL-method}
\title{Remove Subobjects and Cell-Level Meta Data}
\usage{
\S4method{[[}{Seurat,character,missing,`NULL`}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \code{\link{Seurat}} object}

\item{i}{Name(s) of subobject(s) or cell-level meta data to remove}

\item{j}{Ignored}

\item{...}{Arguments passed to \code{\link{LayerData}}}

\item{value}{NULL}
}
\value{
\code{x} with \code{i} removed from the object
}
\description{
Remove Subobjects and Cell-Level Meta Data
}
\seealso{
See \link[=[[.Seurat]{here} for pulling subobjects using \code{[[},
\link[=$.Seurat]{here} for adding metadata with \code{[[<-}, and
\link[=[[<-,Seurat]{here} for adding subobjects with \code{[[<-}

Seurat object, validity, and interaction methods
\code{\link{$.Seurat}()},
\code{\link{Seurat-class}},
\code{\link{Seurat-validity}},
\code{\link{[[.Seurat}()},
\code{\link{[[<-,Seurat}},
\code{\link{dim.Seurat}()},
\code{\link{dimnames.Seurat}()},
\code{\link{merge.Seurat}()},
\code{\link{names.Seurat}()},
\code{\link{subset.Seurat}()}
}
\concept{seurat}
