% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{[[<-,Seurat}
\alias{[[<-,Seurat}
\alias{[[<-.Seurat}
\alias{[[<-,Seurat,character,missing,Assay5-method}
\alias{[[<-,Seurat,character,missing,DimReduc-method}
\alias{[[<-,Seurat,character,missing,Graph-method}
\alias{[[<-,Seurat,character,missing,Neighbor-method}
\alias{[[<-,Seurat,character,missing,SeuratCommand-method}
\alias{[[<-,Seurat,character,missing,SpatialImage-method}
\title{Add Subobjects}
\usage{
\S4method{[[}{Seurat,character,missing,Assay}(x, i, j, ...) <- value

\S4method{[[}{Seurat,character,missing,Assay5}(x, i, j, ...) <- value

\S4method{[[}{Seurat,character,missing,DimReduc}(x, i, j, ...) <- value

\S4method{[[}{Seurat,character,missing,Graph}(x, i, j, ...) <- value

\S4method{[[}{Seurat,character,missing,Neighbor}(x, i, j, ...) <- value

\S4method{[[}{Seurat,character,missing,SeuratCommand}(x, i, j, ...) <- value

\S4method{[[}{Seurat,character,missing,SpatialImage}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \code{\link{Seurat}} object}

\item{i}{Name to add subobject as}

\item{j}{Ignored}

\item{...}{Ignored}

\item{value}{A valid subobject (eg. a \link[Assay]{v3} or
\link[Assay5]{v5} assay or a \link[DimReduc]{dimensional reduction})}
}
\value{
\code{x} with \code{value} added as \code{i}
}
\description{
Add Subobjects

#' @rdname cash-.Seurat
#'
setMethod(
  f = '[[<-',
  signature = c(
    x = 'Seurat',
    i = 'character',
    j = 'missing',
    value = 'data.frame'
  ),
  definition = function(x, i, ..., value) {
    # Because R is stupid sometimes
    if (!length(x = i) && !ncol(x = value)) {
      return(x)
    }
    # Check that the `i` we're adding are present in the data frame
    if (!is.null(x = names(x = value))) {
      i <- match.arg(arg = i, choices = names(x = value), several.ok = TRUE)
    } else if (length(x = i) != ncol(x = value)) {
      stop(
        "Cannot assign ",
        length(x = i),
        " names to ",
        ncol(x = value),
        " bits of meta data",
        call. = FALSE
      )
    }
    # Handle meta data for different cells
    names.intersect <- intersect(x = row.names(x = value), y = colnames(x = x))
    if (length(x = names.intersect)) {
      value <- value[names.intersect, , drop = FALSE]
      if (!nrow(x = value)) {
        stop(
          "None of the cells provided are in this Seurat object",
          call. = FALSE
        )
      }
    } else if (nrow(x = value) == ncol(x = x)) {
      # When no cell names are provided in value, assume it's in cell order
      row.names(x = value) <- colnames(x = x)
    } else {
      # Throw an error when no cell names provided and cannot assume cell order
      stop(
        "Cannot add more or less meta data without cell names",
        call. = FALSE
      )
    }
    # Add the cell-level meta data using the `value = vector` method
    for (n in i) {
      v <- value[[n]]
      names(x = v) <- row.names(x = value)
      x[[n]] <- v
    }
    return(x)
  }
)
}
\seealso{
See \link[=[[.Seurat]{here} for pulling subobjects using \code{[[},
\link[=$.Seurat]{here} for adding metadata with \code{[[<-}, and
\link[=[[<-,Seurat,NULL]{here} for removing subobjects and cell-level meta
data with \code{[[<-}

Seurat object, validity, and interaction methods
\code{\link{$.Seurat}()},
\code{\link{Seurat-class}},
\code{\link{Seurat-validity}},
\code{\link{[[.Seurat}()},
\code{\link{[[<-,Seurat,NULL}},
\code{\link{dim.Seurat}()},
\code{\link{dimnames.Seurat}()},
\code{\link{merge.Seurat}()},
\code{\link{names.Seurat}()},
\code{\link{subset.Seurat}()}
}
\concept{seurat}
