% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{subset.StdAssay}
\alias{subset.StdAssay}
\title{Subset an Assay}
\usage{
\method{subset}{StdAssay}(x, cells = NULL, features = NULL, layers = NULL, ...)
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{cells}{Cell names}

\item{features}{Feature names}

\item{layers}{Layer to keep; defaults to all layers}

\item{...}{Arguments passed to \code{\link{LayerData}}}
}
\value{
\code{x} with just the cells and features specified by
\code{cells} and \code{features} for the layers specified by \code{layers}
}
\description{
Subset an Assay
}
\seealso{
Other stdassay: 
\code{\link{StdAssay-class}},
\code{\link{StdAssay-validity}},
\code{\link{[.StdAssay}()},
\code{\link{[[.StdAssay}()},
\code{\link{dim.StdAssay}()},
\code{\link{dimnames.StdAssay}()},
\code{\link{show,StdAssay-method}}
}
\concept{stdassay}
\keyword{internal}
