% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{SubsetVST}
\alias{SubsetVST}
\title{Subset cells in vst data}
\usage{
SubsetVST(sct.info, cells, features)
}
\arguments{
\item{sct.info}{A vst.out list}

\item{cells}{vector of cells to retain}

\item{features}{vector of features to retain}
}
\description{
Subset cells in vst data
}
\keyword{internal}
