% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Tool}
\alias{Tool}
\alias{Tools}
\alias{Tool<-}
\title{Get and set additional tool data}
\usage{
Tool(object, ...)

Tool(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Information to be added to tool list}
}
\value{
If no additional arguments, returns the names of the tools in the
object; otherwise returns the data placed by the tool requested
}
\description{
Use \code{Tool} to get tool data. If no additional arguments are provided,
will return a vector with the names of tools in the object.
}
\note{
For developers: set tool data using \code{Tool<-}. \code{Tool<-} will
automatically set the name of the tool to the function that called \code{Tool<-},
so each function gets one entry in the tools list and cannot overwrite another
function's entry. The automatic naming will also remove any method identifiers
(eg. RunPCA.Seurat will become RunPCA); please plan accordingly.
}
\concept{data-access}
