% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R
\name{TopN}
\alias{TopN}
\alias{TopCells}
\alias{TopFeatures}
\title{Top Cells and Features}
\usage{
TopCells(object, dim = 1, ncells = 20, balanced = FALSE, ...)

TopFeatures(
  object,
  dim = 1,
  nfeatures = 20,
  projected = FALSE,
  balanced = FALSE,
  ...
)
}
\arguments{
\item{object}{DimReduc object}

\item{dim}{Dimension to use}

\item{balanced}{Return an equal number of cells with both + and - scores.}

\item{...}{Extra parameters passed to \code{\link{Embeddings}} or
\code{\link{Loadings}}}

\item{projected}{Use the projected feature loadings}

\item{ncells/nfeatures}{Number of cells and features to return}
}
\value{
Returns a vector of cells or features
}
\description{
Find cells and features with highest scores for a given dimensional
reduction technique
}
\examples{
pbmc_small
head(TopCells(object = pbmc_small[["pca"]]))
# Can specify which dimension and how many cells to return
TopCells(object = pbmc_small[["pca"]], dim = 2, ncells = 5)

pbmc_small
TopFeatures(object = pbmc_small[["pca"]], dim = 1)
# After projection:
TopFeatures(object = pbmc_small[["pca"]], dim = 1,  projected = TRUE)

}
\concept{dimreduc}
