% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{UpdateAssay}
\alias{UpdateAssay}
\title{Update Seurat assay}
\usage{
UpdateAssay(old.assay, assay)
}
\arguments{
\item{old.assay}{Seurat2 assay}

\item{assay}{Name to store for assay in new object}
}
\description{
Update Seurat assay
}
\keyword{internal}
