% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{UpdateDimReduction}
\alias{UpdateDimReduction}
\title{Update dimension reduction}
\usage{
UpdateDimReduction(old.dr, assay)
}
\arguments{
\item{old.dr}{Seurat2 dimension reduction slot}

\item{assay.used}{Name of assay used to compute dimension reduction}
}
\description{
Update dimension reduction
}
\keyword{internal}
