% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UpdateKey}
\alias{UpdateKey}
\title{Update a Key}
\usage{
UpdateKey(key)
}
\arguments{
\item{key}{A character to become a Seurat Key}
}
\value{
An updated Key that's valid for Seurat
}
\description{
Update a Key
}
\section{\code{Seurat} Object Keys}{

blah
}

\keyword{internal}
