% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{V2Update}
\alias{V2Update}
\alias{UpdateAssay}
\alias{UpdateDimReduction}
\alias{UpdateJackstraw}
\title{Update Seurat v2 Internal Objects}
\usage{
UpdateAssay(old.assay, assay)

UpdateDimReduction(old.dr, assay)

UpdateJackstraw(old.jackstraw)
}
\arguments{
\item{old.assay, old.dr, old.jackstraw}{Seurat v2 assay, dimensional
reduction, or jackstraw object}

\item{assay}{Name to store for assay in new object}

\item{assay.used}{Name of assay used to compute dimension reduction}
}
\value{
A v3/v4 \code{\link{Assay}}, \code{\link{DimReduc}}, or
\code{\link{JackStrawData}} object
}
\description{
Helper functions to update old Seurat v2 objects to v3/v4 objects
}
\keyword{internal}
