% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{v5-assay-summaries}
\alias{v5-assay-summaries}
\alias{colMeans,StdAssay-method}
\alias{colSums,StdAssay-method}
\alias{rowMeans,StdAssay-method}
\alias{rowSums,StdAssay-method}
\title{V5 Assay Summaries}
\usage{
\S4method{colMeans}{StdAssay}(x, na.rm = FALSE, dims = 1, layer = NULL, ...)

\S4method{colSums}{StdAssay}(x, na.rm = FALSE, dims = 1, layer = NULL, ...)

\S4method{rowMeans}{StdAssay}(x, na.rm = FALSE, dims = 1, layer = NULL, ...)

\S4method{rowSums}{StdAssay}(x, na.rm = FALSE, dims = 1, layer = NULL, ...)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric,
    complex, integer or logical values, or a numeric data frame.  For
    \code{.colSums()} etc, a numeric, integer or logical matrix (or
    vector of length \code{m * n}).}

\item{na.rm}{logical.  Should missing values (including \code{NaN})
    be omitted from the calculations?}

\item{dims}{integer number: Which dimensions are regarded as \sQuote{rows} or
    \sQuote{columns} to sum over.  For \code{row*}, the sum or mean is
    over dimensions \code{dims+1, \dots}; for \code{col*} it is over
    dimensions \code{1:dims}.}

\item{layer}{Name of layer to run function on}

\item{...}{Ignored}
}
\value{
The results of the summary math function for the layer specified
}
\description{
Summary maths for \code{\link{StdAssay}} Objects
}
\keyword{internal}
