% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ValidateDataForMerge}
\alias{ValidateDataForMerge}
\title{Validate Assay Data for Merge}
\usage{
ValidateDataForMerge(assay, slot)
}
\arguments{
\item{assay}{Assay to pull data from}

\item{slot}{Slot to pull from}
}
\value{
Returns the data matrix if present (i.e.) not 0x0. Otherwise, returns an
appropriately sized empty sparse matrix
}
\description{
Pulls the proper data matrix for merging assay data. If the slot is empty, will return an empty
matrix with the proper dimensions from one of the remaining data slots.
}
\keyword{internal}
