% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{VariableFeatures-StdAssay}
\alias{VariableFeatures-StdAssay}
\alias{HVFInfo.StdAssay}
\alias{VariableFeatures.StdAssay}
\alias{VariableFeatures<-.StdAssay}
\title{Highly Variable Features}
\usage{
\method{HVFInfo}{StdAssay}(object, method = NULL, status = FALSE, layer = NULL, strip = TRUE, ...)

\method{VariableFeatures}{StdAssay}(object, method = NULL, layer = NULL, collapse = TRUE, ...)

\method{VariableFeatures}{StdAssay}(object, method = "custom", layer = NULL, ...) <- value
}
\arguments{
\item{object}{An object}

\item{status}{Add variable status to the resulting data frame}

\item{...}{Arguments passed to other methods}

\item{value}{A character vector of variable features}
}
\value{
\code{HVFInfo}: A data frame with feature means, dispersion, and
scaled dispersion

\code{VariableFeatures}: a vector of the variable features

\code{SVFInfo}: a data frame with the spatially variable features

\code{SpatiallyVariableFeatures}: a character vector of the spatially
variable features
}
\description{
Get and set variable feature information for an \code{\link{Assay}} object.
\code{HVFInfo} and \code{VariableFeatures} utilize generally variable
features, while \code{SVFInfo} and \code{SpatiallyVariableFeatures} are
restricted to spatially variable features
}
\keyword{internal}
