% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R
\name{VariableFeatures}
\alias{VariableFeatures}
\alias{VariableFeatures<-}
\alias{VariableFeatures.Assay}
\alias{VariableFeatures<-.Assay}
\title{Get and set variable feature information}
\usage{
VariableFeatures(object, ...)

VariableFeatures(object, ...) <- value

\method{VariableFeatures}{Assay}(object, selection.method = NULL, ...)

\method{VariableFeatures}{Assay}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{A character vector of variable features}

\item{selection.method}{Method used to set variable features}
}
\description{
Get and set variable feature information
}
\concept{data-access}
