% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/seurat.R
\name{Version}
\alias{Version}
\alias{Version.Seurat}
\title{Get Version Information}
\usage{
Version(object, ...)

\method{Version}{Seurat}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get Version Information
}
\examples{
Version(pbmc_small)

}
\concept{data-access}
