% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{as.Centroids}
\alias{as.Centroids}
\alias{as.Segmentation}
\alias{as.Centroids.Segmentation}
\alias{as.Segmentation.Centroids}
\title{Convert Segmentation Layers}
\usage{
as.Centroids(x, nsides = NULL, radius = NULL, theta = NULL, ...)

as.Segmentation(x, ...)

\method{as.Centroids}{Segmentation}(x, nsides = NULL, radius = NULL, theta = NULL, ...)

\method{as.Segmentation}{Centroids}(x, ...)
}
\arguments{
\item{x}{An object}

\item{nsides}{The number of sides to represent cells/spots; pass
\code{\link[base]{Inf}} to plot as circles}

\item{radius}{Radius of shapes when plotting}

\item{theta}{Angle to adjust shapes when plotting}

\item{...}{Arguments passed to other methods}
}
\value{
\code{as.Centroids}: A
\code{\link[SeuratObject:Centroids-class]{Centroids}} object

\code{as.Segmentation}: A
\code{\link[SeuratObject:Segmentation-class]{Segmentation}} object
}
\description{
Convert Segmentation Layers
}
