% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{as.sparse}
\alias{as.sparse}
\alias{as.sparse.data.frame}
\alias{as.sparse.Matrix}
\alias{as.sparse.matrix}
\alias{as.sparse.ngCMatrix}
\title{Cast to Sparse}
\usage{
as.sparse(x, ...)

\method{as.sparse}{data.frame}(x, row.names = NULL, ...)

\method{as.sparse}{Matrix}(x, ...)

\method{as.sparse}{matrix}(x, ...)

\method{as.sparse}{ngCMatrix}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to other methods}

\item{row.names}{\code{NULL} or a character vector giving the row names for
the data; missing values are not allowed}
}
\value{
A sparse representation of the input data
}
\description{
Convert dense objects to sparse representations
}
\concept{utils}
