% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\docType{class}
\name{Assay5-class}
\alias{Assay5-class}
\alias{Assay5}
\title{The v5 \code{Assay} Object}
\description{
The v5 \code{Assay} is the typical \code{Assay} class used in \pkg{Seurat}
v5; ...
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{\link[base]{matrix}} or
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} with single-cell expression
data}
}}

\seealso{
\code{\link{StdAssay-class}}
}
