% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AttachDeps}
\alias{AttachDeps}
\title{Attach Required Packages}
\usage{
AttachDeps(deps)
}
\arguments{
\item{deps}{A character vector of packages to attach}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Helper function to attach required packages. Detects if a package is already
attached and if so, skips it. Should be called in \code{\link[base]{.onAttach}}
}
\concept{utils}
