% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/default.R, R/dimreduc.R,
%   R/neighbor.R
\name{Cells}
\alias{Cells}
\alias{Features}
\alias{Cells.default}
\alias{Cells.DimReduc}
\alias{Cells.Neighbor}
\title{Cell and Feature Names}
\usage{
Cells(x, ...)

Features(x, ...)

\method{Cells}{default}(x, ...)

\method{Cells}{DimReduc}(x, ...)

\method{Cells}{Neighbor}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{Cell}: A vector of cell names

\code{Features}: A vector of feature names
}
\description{
Get the cell and feature names of an object
}
\examples{
Cells(x = pbmc_small)

}
\concept{data-access}
