% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Cells}
\alias{Cells}
\title{Get cells present in an object}
\usage{
Cells(x)
}
\arguments{
\item{x}{An object}
}
\value{
A vector of cell names
}
\description{
Get cells present in an object
}
\examples{
Cells(x = pbmc_small)

}
