% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimreduc.R, R/spatial.R
\name{Cells}
\alias{Cells}
\alias{Cells.DimReduc}
\alias{Cells.SlideSeq}
\alias{Cells.STARmap}
\alias{Cells.VisiumV1}
\title{Get cells present in an object}
\usage{
Cells(x)

\method{Cells}{DimReduc}(x)

\method{Cells}{SlideSeq}(x)

\method{Cells}{STARmap}(x)

\method{Cells}{VisiumV1}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A vector of cell names
}
\description{
Get cells present in an object
}
\examples{
Cells(x = pbmc_small)

}
