% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay5.R
\name{CreateAssay5Object}
\alias{CreateAssay5Object}
\alias{CreateAssay5Object.default}
\alias{CreateAssay5Object.list}
\alias{CreateAssay5Object.Matrix}
\alias{CreateAssay5Object.matrix}
\alias{CreateAssay5Object.spam}
\title{Create a v5 Assay object}
\usage{
CreateAssay5Object(
  counts,
  min.cells = 0,
  min.features = 0,
  transpose = FALSE,
  ...
)

\method{CreateAssay5Object}{default}(
  counts,
  min.cells = 0,
  min.features = 0,
  layer = "counts",
  transpose = FALSE,
  csum = NULL,
  fsum = NULL,
  ...
)

\method{CreateAssay5Object}{list}(
  counts,
  min.cells = 0,
  min.features = 0,
  transpose = FALSE,
  csum = NULL,
  fsum = NULL,
  ...
)

\method{CreateAssay5Object}{Matrix}(
  counts,
  min.cells = 0,
  min.features = 0,
  transpose = FALSE,
  layer = "counts",
  ...
)

\method{CreateAssay5Object}{matrix}(
  counts,
  min.cells = 0,
  min.features = 0,
  transpose = FALSE,
  layer = "counts",
  ...
)

\method{CreateAssay5Object}{spam}(
  counts,
  min.cells = 0,
  min.features = 0,
  transpose = FALSE,
  layer = "counts",
  ...
)
}
\arguments{
\item{counts}{A two-dimensional expression matrix}

\item{min.cells}{Include features detected in at least this many cells;
will subset the counts matrix as well. To reintroduce excluded features,
create a new object with a lower cutoff}

\item{min.features}{Include cells where at least this many features
are detected}

\item{transpose}{Create a transposed assay}

\item{...}{Arguments passed on to other methods}

\item{layer}{Name of layer to store \code{counts} as}

\item{csum}{Function for calculating cell sums}

\item{fsum}{Function for calculating feature sums}
}
\value{
An \code{\link{Assay5}} object
}
\description{
Create an \code{\link{Assay5}} object from a feature expression matrix;
the expected format of the matrix is features x cells
}
