% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/segmentation.R
\name{CreateSegmentation}
\alias{CreateSegmentation}
\alias{CreateSegmentation.data.frame}
\alias{CreateSegmentation.Segmentation}
\alias{CreateSegmentation.sf}
\title{Create a \code{\link[SeuratObject:Segmentation-class]{Segmentation}} Objects}
\usage{
CreateSegmentation(coords, compact = FALSE)

\method{CreateSegmentation}{data.frame}(coords, compact = FALSE)

\method{CreateSegmentation}{Segmentation}(coords, compact = FALSE)

\method{CreateSegmentation}{sf}(coords, compact = TRUE)
}
\arguments{
\item{coords}{The coordinates of cell segmentations}

\item{compact}{Logical indicating whether or not the object should only store segmentation data
in the \code{sf.data} slot; see \link{Segmentation-class} for details.}
}
\value{
A \code{\link[SeuratObject:Segmentation-class]{Segmentation}} object
}
\description{
Create a \code{\link[SeuratObject:Segmentation-class]{Segmentation}} Objects
}
\concept{spatial}
