% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatialcoords.R
\name{Crop}
\alias{Crop}
\alias{Crop.SpatialCoords}
\title{Crop Coordinates}
\usage{
Crop(object, x = NULL, y = NULL, coords = c("plot", "tissue"), ...)

\method{Crop}{SpatialCoords}(object, x = NULL, y = NULL, coords = c("plot", "tissue"), ...)
}
\arguments{
\item{object}{An object}

\item{x, y}{Range to crop x/y limits to; if \code{NULL}, uses full range of
\code{x}/\code{y}}

\item{coords}{Coordinate system to execute crop; choose from:
\itemize{
 \item \dQuote{\code{plot}}: Coordinates as shown when plotting
 \item \dQuote{\code{tissue}}: Coordinates from
  \code{\link{GetTissueCoordinates}}
}}

\item{...}{...}
}
\value{
\code{object} cropped to the region specified by \code{x}
and \code{y}
}
\description{
Crop Coordinates
}
