% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R, R/graph.R
\name{DefaultAssay}
\alias{DefaultAssay}
\alias{DefaultAssay<-}
\alias{DefaultAssay.Assay}
\alias{DefaultAssay<-.Assay}
\alias{DefaultAssay.Graph}
\alias{DefaultAssay<-.Graph}
\title{Default Assay}
\usage{
DefaultAssay(object, ...)

DefaultAssay(object, ...) <- value

\method{DefaultAssay}{Assay}(object, ...)

\method{DefaultAssay}{Assay}(object, ...) <- value

\method{DefaultAssay}{Graph}(object, ...)

\method{DefaultAssay}{Graph}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Name of assay to set as default}
}
\value{
\code{DefaultAssay}: The name of the default assay

\code{DefaultAssay<-}: An object with the default assay updated
}
\description{
Get and set the default assay
}
