% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{.DiskLoad}
\alias{.DiskLoad}
\alias{.DiskLoad.default}
\alias{.DiskLoad.H5ADMatrix}
\alias{.DiskLoad.HDF5Matrix}
\title{Disk Loading Function}
\usage{
.DiskLoad(x)

\method{.DiskLoad}{default}(x)

\method{.DiskLoad}{H5ADMatrix}(x)

\method{.DiskLoad}{HDF5Matrix}(x)
}
\arguments{
\item{x}{A file-backed object}
}
\value{
A one-length character that defines a function to load a matrix from
a file
}
\description{
Generate a function to load a matrix from an on-disk file
}
\keyword{internal}
