% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.GetMethod}
\alias{.GetMethod}
\title{Get a Method}
\usage{
.GetMethod(fxn, cls)
}
\arguments{
\item{fxn}{Name of a function as a character}

\item{cls}{The class to find a method of \code{fxn} for}
}
\value{
The method of \code{fxn} for class \code{cls}; if no method found,
returns the default method. If no default method found; returns \code{NULL}
}
\description{
Get a Method
}
\examples{
.GetMethod('t', 'Matrix')
.GetMethod('t', 'data.frame')

}
\concept{utils}
\keyword{internal}
