% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimreduc.R, R/seurat.R
\name{Embeddings}
\alias{Embeddings}
\alias{Embeddings.DimReduc}
\alias{Embeddings.Seurat}
\alias{Embeddings.Seurat5}
\title{Get Cell Embeddings}
\usage{
Embeddings(object, ...)

\method{Embeddings}{DimReduc}(object, ...)

\method{Embeddings}{Seurat}(object, reduction = "pca", ...)

\method{Embeddings}{Seurat5}(object, reduction = "pca", ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{reduction}{Name of reduction to pull cell embeddings for}
}
\value{
The embeddings matrix
}
\description{
Get Cell Embeddings
}
\examples{
# Get the embeddings directly from a DimReduc object
Embeddings(object = pbmc_small[["pca"]])[1:5, 1:5]

# Get the embeddings from a specific DimReduc in a Seurat object
Embeddings(object = pbmc_small, reduction = "pca")[1:5, 1:5]

}
\concept{data-access}
