% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{FindObject}
\alias{FindObject}
\title{Find the collection of an object within a Seurat object}
\usage{
FindObject(object, name)
}
\arguments{
\item{object}{A Seurat object}

\item{name}{Name of object to find}
}
\value{
The collection (slot) of the object
}
\description{
Find the collection of an object within a Seurat object
}
\keyword{internal}
