% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/seurat.R
\name{GetTissueCoordinates}
\alias{GetTissueCoordinates}
\alias{GetTissueCoordinates.Seurat}
\title{Get tissue coordinates}
\usage{
GetTissueCoordinates(object, ...)

\method{GetTissueCoordinates}{Seurat}(object, image = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{image}{Name of \code{SpatialImage} object to get coordinates for; if
\code{NULL}, will attempt to select an image automatically}
}
\value{
A data frame with tissue coordinates
}
\description{
Get tissue coordinates
}
\concept{data-access}
