% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/fov.R, R/seurat.R
\name{GetTissueCoordinates}
\alias{GetTissueCoordinates}
\alias{GetTissueCoordinates.FOV}
\alias{GetTissueCoordinates.Seurat}
\title{Get tissue coordinates}
\usage{
GetTissueCoordinates(object, ...)

\method{GetTissueCoordinates}{FOV}(object, which = NULL, ...)

\method{GetTissueCoordinates}{Seurat}(object, image = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{which}{Name of segmentation boundary or molecule set to retrieve coordinates for}

\item{image}{Name of \code{SpatialImage} object to get coordinates for; if
\code{NULL}, will retrieve coordinates from the default image}
}
\value{
A data frame with tissue coordinates

\code{GetTissueCoordinates}: ...
}
\description{
Get tissue coordinates
}
\examples{
\dontrun{
GetTissueCoordinates(object, which = "centroids")
}
}
\concept{data-access}
