% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsMatrixEmpty}
\alias{IsMatrixEmpty}
\title{Check if a matrix is empty}
\usage{
IsMatrixEmpty(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Whether or not \code{x} is empty
}
\description{
Takes a matrix and asks if it's empty (either 0x0 or 1x1 with a value of NA)
}
\examples{
IsMatrixEmpty(new("matrix"))
IsMatrixEmpty(matrix())
IsMatrixEmpty(matrix(1:3))

}
\concept{utils}
