% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{JS}
\alias{JS}
\alias{JS<-}
\title{Get and set JackStraw information}
\usage{
JS(object, ...)

JS(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{JackStraw information}
}
\description{
Get and set JackStraw information
}
