% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay5.R
\name{LayerData}
\alias{LayerData}
\alias{LayerData<-}
\alias{Layers}
\alias{LayerData.StdAssay}
\alias{LayerData<-.StdAssay}
\alias{Layers.StdAssay}
\title{Query and Manipulate Assay Layers}
\usage{
LayerData(object, layer, ...)

LayerData(object, layer, ...) <- value

Layers(object, ...)

\method{LayerData}{StdAssay}(object, layer = NULL, fast = FALSE, ...)

\method{LayerData}{StdAssay}(object, layer, features = NULL, cells = NULL, ...) <- value

\method{Layers}{StdAssay}(object, ...)
}
\arguments{
\item{object}{An object}

\item{layer}{Name of layer to fetch or set}

\item{...}{Arguments passed to other methods}

\item{value}{New two-dimensional data to be added as a layer}

\item{fast}{Determine how to return the layer data; choose from:
\describe{
 \item{\code{FALSE}}{Apply any transpositions and attempt to add feature/cell
  names (if supported) back to the layer data}
 \item{\code{NA}}{Attempt to add feature/cell names back to the layer data,
  skip any transpositions}
 \item{\code{TRUE}}{Do not apply any transpositions or add feature/cell names
  to the layer data}
}}

\item{features, cells}{Vectors of features/cells to include ...}
}
\value{
\code{LayerData}: the layer data for \code{layer} from \code{object}

\code{Layer<-}: \code{object} with \code{value} added as a layer
named \code{layer}

\code{Layers}: the names of the layers present in \code{object}
}
\description{
Query and Manipulate Assay Layers
}
\concept{data-access}
