% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Loadings}
\alias{Loadings}
\alias{Loadings<-}
\title{Get and set feature loadings}
\usage{
Loadings(object, ...)

Loadings(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Feature loadings to add}
}
\description{
Get and set feature loadings
}
