% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/default.R
\name{MatchCells}
\alias{MatchCells}
\alias{MatchCells.character}
\alias{MatchCells.numeric}
\title{Match Cells}
\usage{
MatchCells(new, orig, ordered = FALSE)

\method{MatchCells}{character}(new, orig, ordered = FALSE)

\method{MatchCells}{numeric}(new, orig, ordered = FALSE)
}
\arguments{
\item{new}{A vector of new cells}

\item{orig}{A vector of existing cells}

\item{ordered}{Sort the result to the same order as \code{orig}}
}
\value{
A numeric vector with new cells in order of the original cells
}
\description{
Match Cells
}
