% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R, R/dimreduc.R, R/seurat.R
\name{Misc}
\alias{Misc}
\alias{Misc<-}
\alias{Misc.StdAssay}
\alias{Misc<-.StdAssay}
\alias{Misc.DimReduc}
\alias{Misc<-.DimReduc}
\alias{Misc.Seurat}
\alias{Misc<-.Seurat}
\title{Get and set miscellaneous data}
\usage{
Misc(object, ...)

Misc(object, ...) <- value

\method{Misc}{StdAssay}(object, slot = NULL, ...)

\method{Misc}{StdAssay}(object, slot, ...) <- value

\method{Misc}{DimReduc}(object, slot = NULL, ...)

\method{Misc}{DimReduc}(object, slot, ...) <- value

\method{Misc}{Seurat}(object, slot = NULL, ...)

\method{Misc}{Seurat}(object, slot, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Data to add}

\item{slot}{Name of specific bit of meta data to pull}
}
\value{
Miscellaneous data

An object with miscellaneous data added
}
\description{
Get and set miscellaneous data
}
\examples{
# Get the misc info
Misc(object = pbmc_small, slot = "example")

# Add misc info
Misc(object = pbmc_small, slot = "example") <- "testing_misc"

}
\concept{data-access}
