% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{NullImage}
\alias{NullImage}
\title{Return a null image}
\usage{
NullImage(mode)
}
\arguments{
\item{mode}{Image representation to return
see \code{\link{GetImage}} for more details}
}
\description{
Return a null image
}
\keyword{internal}
