% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PackageCheck}
\alias{PackageCheck}
\title{Check the existence of a package}
\usage{
PackageCheck(..., error = TRUE)
}
\arguments{
\item{...}{Package names}

\item{error}{If true, throw an error if the package doesn't exist}
}
\value{
Invisibly returns boolean denoting if the package is installed
}
\description{
Check the existence of a package
}
\section{Lifecycle}{


\Sexpr[stage=build,results=rd]{lifecycle::badge("deprecated")}

\code{PackageCheck} was deprecated in version 5.0.0; please use
\code{\link[rlang:check_installed]{rlang::check_installed}()} instead
}

\concept{utils}
