% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RandomName}
\alias{RandomName}
\title{Generate a random name}
\usage{
RandomName(length = 5L, ...)
}
\arguments{
\item{length}{How long should the name be}

\item{...}{Extra parameters passed to \code{\link[base]{sample}}}
}
\value{
A character with \code{nchar == length} of randomly sampled letters
}
\description{
Make a name from randomly sampled lowercase letters, pasted together with no
spaces or other characters
}
\examples{
\dontrun{
set.seed(42L)
RandomName()
RandomName(7L, replace = TRUE)
}

}
\seealso{
\code{\link[base]{sample}}
}
\keyword{internal}
