% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plan}
\alias{colMeans}
\alias{handlers}
\alias{with_progress}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{future}{\code{\link[future]{plan}}}

  \item{Matrix}{\code{\link[Matrix:colSums]{colMeans}}}

  \item{progressr}{\code{\link[progressr]{handlers}}, \code{\link[progressr]{with_progress}}}
}}

