% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R, R/dimreduc.R,
%   R/spatial.R
\name{RenameCells}
\alias{RenameCells}
\alias{RenameCells.Assay}
\alias{RenameCells.DimReduc}
\alias{RenameCells.VisiumV1}
\title{Rename cells}
\usage{
RenameCells(object, ...)

\method{RenameCells}{Assay}(object, new.names = NULL, ...)

\method{RenameCells}{DimReduc}(object, new.names = NULL, ...)

\method{RenameCells}{VisiumV1}(object, new.names = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{new.names}{vector of new cell names}
}
\value{
An object with new cell names
}
\description{
Change the cell names in all the different parts of an object. Can be useful
before combining multiple objects.
}
\examples{
# Rename cells in an Assay
head(x = colnames(x = pbmc_small[["RNA"]]))
renamed.assay <- RenameCells(
    object = pbmc_small[["RNA"]],
    new.names = paste0("A_", colnames(x = pbmc_small[["RNA"]]))
)
head(x = colnames(x = renamed.assay))

# Rename cells in a DimReduc
head(x = Cells(x = pbmc_small[["pca"]]))
renamed.dimreduc <- RenameCells(
    object = pbmc_small[["pca"]],
    new.names = paste0("A_", Cells(x = pbmc_small[["pca"]]))
)
head(x = Cells(x = renamed.dimreduc))

}
\concept{seurat}
