% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{SaveSeuratRds}
\alias{SaveSeuratRds}
\title{Save Seurat Objects to RDS files}
\usage{
SaveSeuratRds(object, file = NULL, destdir = NULL, relative = FALSE, ...)
}
\arguments{
\item{object}{A \code{\link{Seurat}} object}

\item{file}{Path to save \code{object} to; defaults to
\code{file.path(getwd(), paste0(Project(object), ".Rds"))}}

\item{destdir}{Destination directory for on-disk layers saved in
\dQuote{\code{\Sexpr[stage=render]{tempdir()}}}}

\item{relative}{Save relative paths instead of absolute ones}

\item{...}{
  Arguments passed on to \code{\link[base:readRDS]{base::saveRDS}}
  \describe{
    \item{\code{ascii}}{a logical.  If \code{TRUE} or \code{NA}, an ASCII
    representation is written; otherwise (default), a binary one is used.
    See the comments in the help for \code{\link[base]{save}}.}
    \item{\code{version}}{the workspace format version to use.  \code{NULL}
    specifies the current default version (3). The only other supported
    value is 2, the default from \R 1.4.0 to \R 3.5.0.}
    \item{\code{compress}}{a logical specifying whether saving to a named file is
    to use \code{"gzip"} compression, or one of \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} to indicate the type of compression to
    be used.  Ignored if \code{file} is a connection.}
    \item{\code{refhook}}{a hook function for handling reference objects.}
  }}
}
\value{
Invisibly returns \code{file}
}
\description{
Save Seurat Objects to RDS files
}
\section{Progress Updates with \pkg{progressr}}{

This function uses
\href{https://cran.r-project.org/package=progressr}{\pkg{progressr}} to
render status updates and progress bars. To enable progress updates, wrap
the function call in \code{\link[progressr]{with_progress}} or run
\code{\link[progressr:handlers]{handlers(global = TRUE)}} before running
this function. For more details about \pkg{progressr}, please read
\href{https://progressr.futureverse.org/articles/progressr-intro.html}{\code{vignette("progressr-intro")}}
}

\examples{
if (requireNamespace("HDF5Array") && requireNamespace("fs")) {
  out <- tempfile(fileext = ".Rds")
  pbmc_small[["disk"]] <- CreateAssay5Object(list(
    mem = LayerData(pbmc_small, "counts"),
    disk = as(LayerData(pbmc_small, "counts"), "HDF5Array")
  ))
  SaveSeuratRds(pbmc_small, file = out)
  obj <- readRDS(out)
  Tool(obj, "SaveSeuratRds")
}

}
