% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\docType{class}
\name{Segmentation-class}
\alias{Segmentation-class}
\title{The \code{Segmentation} Class}
\description{
A container for cell segmentation boundaries.
Inherits from \code{\link[sp:SpatialPolygons-class]{SpatialPolygons}}.
Supports storing boundaries in \code{\link[sf]{sf}} objects.
}
\section{Slots}{

\describe{
\item{\code{sf.data}}{Segmentation boundaries in \code{\link[sf]{sf}} format}
}}

\seealso{
\code{Segmentation} methods: \code{\link{Segmentation-methods}}

Segmentation layer classes:
\code{\link{Centroids-class}},
\code{\link{Centroids-methods}},
\code{\link{Molecules-class}},
\code{\link{Molecules-methods}},
\code{\link{Segmentation-methods}}
}
\concept{segmentation}
