% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{SeuratCommand-methods}
\alias{SeuratCommand-methods}
\alias{.DollarNames.SeuratCommand}
\alias{$.SeuratCommand}
\alias{[.SeuratCommand}
\alias{show,SeuratCommand-method}
\title{\code{SeuratCommand} Methods}
\usage{
\method{.DollarNames}{SeuratCommand}(x, pattern = "")

\method{$}{SeuratCommand}(x, i, ...)

\method{[}{SeuratCommand}(x, i, ...)

\S4method{show}{SeuratCommand}(object)
}
\arguments{
\item{x, object}{A \code{\link{SeuratCommand}} object}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }

\item{i}{For a \code{$}, a parameter name; for \code{[}, a
\code{SeuratCommand} slot name}

\item{...}{Arguments passed to other methods}
}
\value{
\code{$}: The value for parameter \code{i}

\code{[}: Slot \code{i} from \code{x}

\code{show}: Prints summary to \code{\link[base]{stdout}} and
invisibly returns \code{NULL}
}
\description{
Methods for \code{\link{SeuratCommand}} objects for generics defined in
other packages
}
\section{Functions}{
\itemize{
\item \code{.DollarNames.SeuratCommand}: Autocompletion for \code{$} access on a
\code{SeuratCommand} object

\item \code{$.SeuratCommand}: Access a parameter from a
\code{SeuratCommand} object

\item \code{[.SeuratCommand}: Access data from a \code{SeuratCommand}
object

\item \code{show,SeuratCommand-method}: Overview of a \code{SeuratCommand} object
}}

\concept{command}
