% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{Simplify}
\alias{Simplify}
\alias{Simplify.Spatial}
\title{Simplify Geometry}
\usage{
Simplify(coords, tol, topologyPreserve = TRUE)

\method{Simplify}{Spatial}(coords, tol, topologyPreserve = TRUE)
}
\arguments{
\item{coords}{A `Segmentation` object}

\item{tol}{Numerical tolerance value to be used by the Douglas-Peuker algorithm}

\item{topologyPreserve}{Logical determining if the algorithm should attempt to preserve the topology of the original geometry}
}
\value{
A simplified version of \code{coords}

A `Segmentation` object with simplified segmentation vertices
}
\description{
Simplify Geometry

Simplify segmentations by reducing the number of vertices
}
\concept{spatial}
