% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SparseEmptyMatrix}
\alias{SparseEmptyMatrix}
\title{Generate empty dgC sparse matrix}
\usage{
SparseEmptyMatrix(nrow, ncol, rownames = NULL, colnames = NULL)
}
\description{
Generate empty dgC sparse matrix
}
