% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay.R
\name{SpatiallyVariableFeatures}
\alias{SpatiallyVariableFeatures}
\alias{SpatiallyVariableFeatures.Assay}
\title{Get spatially variable feature information}
\usage{
SpatiallyVariableFeatures(object, ...)

\method{SpatiallyVariableFeatures}{Assay}(
  object,
  selection.method = "markvariogram",
  decreasing = TRUE,
  ...
)
}
\arguments{
\item{decreasing}{Return features in decreasing order (most spatially
variable first).}
}
\description{
Get spatially variable feature information
}
