% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{split.Assay}
\alias{split.Assay}
\title{Split an Assay}
\usage{
\method{split}{Assay}(x, f, drop = FALSE, layers = NA, ret = c("multiassay", "layers"), ...)
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{f}{a \sQuote{factor} in the sense that \code{\link[base]{as.factor}(f)}
    defines the grouping, or a list of such factors in which case their
    interaction is used for the grouping. If \code{x} is a data frame,
    \code{f} can also be a formula of the form \code{ ~ g} to split by
    the variable \code{g}, or more generally of the form \code{ ~ g1 +
      \dots + gk} to split by the interaction of the variables
    \code{g1}, \dots, \code{gk}, where these variables are evaluated in
    the data frame \code{x} using the usual non-standard evaluation
    rules.}

\item{drop}{logical indicating if levels that do not occur should be dropped
    (if \code{f} is a \code{factor} or a list).}

\item{layers}{Names of layers to include in the split; pass \code{NA} for
all layers; pass \code{NULL} for the \link[DefaultLayer]{default layer}}

\item{ret}{Type of return value; choose from:
\itemize{
 \item \dQuote{\code{multiassay}}: a list of \code{\link{Assay}} objects
 \item \dQuote{\code{layers}}: a list of layer matrices
}}

\item{...}{Ignored}
}
\value{
Depends on the value of \code{ret}:
\itemize{
 \item \dQuote{\code{multiassay}}: a list of \code{\link{Assay}} objects;
 the list contains one value per split and each assay contains only the
 layers requested in \code{layers} with the \link[Key]{key} set to the split
 \item \dQuote{\code{layers}}: a list of matrices of length
 \code{length(assays) * length(unique(f))}; the list is named as
 \dQuote{\code{layer.split}}
}
}
\description{
Split an Assay
}
\seealso{
v3 Assay object, validity, and interaction methods
\code{\link{Assay-class}},
\code{\link{Assay-validity}},
\code{\link{CreateAssayObject}()},
\code{\link{[.Assay}()},
\code{\link{dim.Assay}()},
\code{\link{dimnames.Assay}()},
\code{\link{merge.Assay}()},
\code{\link{subset.Assay}()}
}
\concept{assay}
