% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/assay5.R
\name{SplitLayers}
\alias{SplitLayers}
\alias{JoinLayers}
\alias{JoinLayers.Assay5}
\title{Split and Join Layers Together}
\usage{
SplitLayers(object, ...)

JoinLayers(object, ...)

\method{JoinLayers}{Assay5}(object, layers = NULL, new = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{layers}{...}

\item{new}{...}
}
\value{
...
}
\description{
Split and Join Layers Together
}
