% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\docType{class}
\name{StdAssay-class}
\alias{StdAssay-class}
\alias{StdAssay}
\title{Core Assay Infrastructure}
\description{
The \code{StdAssay} class is a virtual class that provides core
infrastructure for assay data in \pkg{Seurat}. Assays contain expression
data along with additional representations of the expression data (layers)
and associated metadata. Derived classes
(eg. \link[Assay-class]{the v5 Assay}) must set the storage mechanism for
expression data (eg. \code{\link[base]{matrix}} or
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}) and may optionally define
additional functionality
}
\section{Slots}{

\describe{
\item{\code{data}}{A two-dimensional object with expression data with cells as
columns and features as rows}

\item{\code{layers}}{A named list containing alternate representations of
\code{data}; layers must have the same cells as \code{data} and either the
same or a subset of the features present in \code{data}}

\item{\code{key}}{...}

\item{\code{cells}}{A vector of cell names present in the assay}

\item{\code{assay.orig}}{...}

\item{\code{features}}{A matrix containing the presence/absence status of features
in the layers with the following properties;
\itemize{
 \item The dimensions of the matrix should be m\emph{f} rows and n\emph{l}
 columns where m\emph{f} is the number of features (rows) in \code{data}
 and n\emph{l} is the number of layers in \code{layers}
 \item The row names should be the features (in order) present in \code{data}
 \item The column names should be the names of each layer present in
 \code{layers}
 \item Each value should be either \code{TRUE} if a given feature is present
 in a given layer, otherwise \code{FALSE}
}}

\item{\code{meta.data}}{...}

\item{\code{misc}}{...}
}}

\seealso{
\code{\link{Assay5-class}}
}
