% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{StdAssay-methods}
\alias{StdAssay-methods}
\alias{[.StdAssay}
\alias{[[.StdAssay}
\alias{dim.StdAssay}
\alias{dimnames.StdAssay}
\alias{dimnames<-.StdAssay}
\alias{head.StdAssay}
\alias{merge.StdAssay}
\alias{subset.StdAssay}
\alias{tail.StdAssay}
\alias{[[<-,StdAssay,character,missing,data.frame-method}
\alias{[[<-,StdAssay,character,missing,factor-method}
\alias{[[<-,StdAssay,character,missing,NULL-method}
\alias{[[<-,StdAssay,character,missing,vector-method}
\alias{[[<-,StdAssay,numeric,missing,ANY-method}
\alias{[[<-,StdAssay,missing,missing,data.frame-method}
\alias{[[<-,StdAssay,missing,missing,NULL-method}
\alias{show,StdAssay-method}
\title{\code{StdAssay} Methods}
\usage{
\method{[}{StdAssay}(x, i, j, ...)

\method{[[}{StdAssay}(x, i, ..., drop = FALSE)

\method{dim}{StdAssay}(x)

\method{dimnames}{StdAssay}(x)

\method{dimnames}{StdAssay}(x) <- value

\method{head}{StdAssay}(x, n = 10L, ...)

\method{merge}{StdAssay}(x, y, ...)

\method{subset}{StdAssay}(x, cells = NULL, features = NULL, ...)

\method{tail}{StdAssay}(x, n = 10L, ...)

\S4method{[[}{StdAssay,character,missing,data.frame}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,character,missing,factor}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,character,missing,`NULL`}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,character,missing,vector}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,numeric,missing,ANY}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,missing,missing,data.frame}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,missing,missing,`NULL`}(x, i, j, ...) <- value

\S4method{show}{StdAssay}(object)
}
\arguments{
\item{x, object}{An \code{\link{StdAssay}} object}

\item{i, features}{For \code{[[}: metadata names; for all other methods,
feature names or indices}

\item{j, cells}{Cell names or indices}

\item{...}{Arguments passed to other methods}

\item{drop}{logical.  If \code{TRUE} the result is coerced to the
    lowest possible dimension.  The default is to drop if only one
    column is left, but \bold{not} to drop if only one row is left.}

\item{value}{For \code{[[} Feature-level metadata to add to the assay; for
\code{dimnames<-}, a list of two character vectors with the first entry being
new feature names and the second being new cell names for the assay}
}
\value{
\code{[}: The \code{data} slot for features \code{i} and cells
\code{j}

\code{[[}: The feature-level metadata for \code{i}

\code{dim}: The number of features (\code{nrow}) and cells
(\code{ncol})

\code{dimnames}: A list with feature (row) and cell (column) names

\code{dimnames<-}: \code{x} with the cell and/or feature
names updated to \code{value}

\code{head}: The first \code{n} rows of feature-level metadata

\code{merge}: A new assay ...

\code{subset}: The assay subsetted to the cells and/or features given

\code{tail}: The last \code{n} rows of feature-level metadata

\code{[[<-}: \code{x} with the metadata updated

\code{show}: Prints summary to \code{\link[base]{stdout}} and
invisibly returns \code{NULL}
}
\description{
Methods for \code{\link{StdAssay}} objects for generics defined in
other packages
}
\details{
The following methods are provided for interacting with a
\code{StdAssay} object

\code{[}: Get expression data from an \code{StdAssay}

\code{[[}: Get feature-level metadata

\code{dim}: Number of cells and features for an \code{StdAssay}

\code{dimnames}: Get the feature and cell names

\code{dimnames<-}: Set the feature and cell names

\code{head} and \code{tail}: Get the first or last rows of feature
level meta data

\code{merge}: Merge multiple assays together

\code{subset}: Subset an assay to a given set of cells
and/or features. \strong{Note}: reordering of cells/features is
\emph{not} permitted

\code{[[<-}: Add or remove pieces of meta data

\code{show}: Overview of an \code{StdAssay} object
}
\seealso{
\code{\link{Cells}} \code{\link{Features}}
}
\concept{assay}
