% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{[.Assay}
\alias{[.Assay}
\title{Get Expression Data}
\usage{
\method{[}{Assay}(x, i, j, ...)
}
\arguments{
\item{x}{An \code{\link{Assay}} object}

\item{i}{Feature names or indices}

\item{j}{Cell names or indices}

\item{...}{Arguments passed to other methods}
}
\description{
Get Expression Data
}
\examples{
rna <- pbmc_small[["RNA"]]
rna[1:10, 1:4]

}
\seealso{
\code{\link{GetAssayData}}

v3 Assay object, validity, and interaction methods
\code{\link{Assay-class}},
\code{\link{Assay-validity}},
\code{\link{CreateAssayObject}()},
\code{\link{[[.Assay}()},
\code{\link{dim.Assay}()},
\code{\link{dimnames.Assay}()},
\code{\link{merge.Assay}()},
\code{\link{subset.Assay}()}
}
\concept{assay}
