% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{[.Assay5}
\alias{[.Assay5}
\title{Get Expression Data}
\usage{
\method{[}{Assay5}(x, i, j, ...)
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{i}{Feature names or indices}

\item{j}{Cell names or indices}

\item{...}{Arguments passed to \code{\link{LayerData}}}
}
\value{
The expression matrix for the default layer of \code{x}
for the \code{i} features and \code{j} cells
}
\description{
Get Expression Data
}
\seealso{
\code{\link{LayerData}}

Other assay5: 
\code{\link{Assay5-class}},
\code{\link{Assay5-validity}},
\code{\link{[[.Assay5}()},
\code{\link{dim.Assay5}()},
\code{\link{dimnames.Assay5}()},
\code{\link{merge.Assay5}()},
\code{\link{subset.Assay5}()}
}
\concept{assay5}
