% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{[.StdAssay}
\alias{[.StdAssay}
\title{Get Expression Data}
\usage{
\method{[}{StdAssay}(x, i, j, ...)
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{i}{Feature names or indices}

\item{j}{Cell names or indices}

\item{...}{Arguments passed to \code{\link{LayerData}}}
}
\value{
The expression matrix for the default layer of \code{x}
for the \code{i} features and \code{j} cells
}
\description{
Get Expression Data
}
\seealso{
Other stdassay: 
\code{\link{StdAssay-class}},
\code{\link{StdAssay-validity}},
\code{\link{[[.StdAssay}()},
\code{\link{dim.StdAssay}()},
\code{\link{dimnames.StdAssay}()},
\code{\link{show,StdAssay-method}},
\code{\link{subset.StdAssay}()}
}
\concept{stdassay}
\keyword{internal}
