% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{[[.StdAssay}
\alias{[[.StdAssay}
\alias{head.StdAssay}
\alias{tail.StdAssay}
\alias{[[<-,StdAssay,character,missing,data.frame-method}
\alias{[[<-,StdAssay,missing,missing,data.frame-method}
\alias{[[<-,StdAssay,character,missing,factor-method}
\alias{[[<-,StdAssay,character,missing,NULL-method}
\alias{[[<-,StdAssay,character,missing,vector-method}
\alias{[[<-,StdAssay,numeric,missing,ANY-method}
\alias{[[<-,StdAssay,missing,missing,NULL-method}
\title{Feature-Level Meta Data}
\usage{
\method{[[}{StdAssay}(x, i, ..., drop = FALSE)

\method{head}{StdAssay}(x, n = 10L, ...)

\method{tail}{StdAssay}(x, n = 10L, ...)

\S4method{[[}{StdAssay,character,missing,data.frame}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,missing,missing,data.frame}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,character,missing,factor}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,character,missing,`NULL`}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,character,missing,vector}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,numeric,missing,ANY}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,missing,missing,data.frame}(x, i, j, ...) <- value

\S4method{[[}{StdAssay,missing,missing,`NULL`}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{i}{Name of feature-level meta data to fetch or add}

\item{...}{Ignored}

\item{drop}{See \code{\link{drop}}}

\item{n}{Number of meta data rows to show}

\item{j}{Ignored}
}
\value{
\code{[[}: The feature-level meta data for \code{i}

\code{[[<-}: \code{x} with \code{value} added as \code{i}
in feature-level meta data

\code{head}: The first \code{n} rows of feature-level meta data

\code{tail}: the last \code{n} rows of feature-level meta data
}
\description{
Get and set feature-level meta data
}
\seealso{
v5 Standard Assay object, validity, and interaction methods
\code{\link{StdAssay-class}},
\code{\link{StdAssay-validity}},
\code{\link{[.StdAssay}()},
\code{\link{dim.StdAssay}()},
\code{\link{dimnames.StdAssay}()},
\code{\link{show,StdAssay-method}},
\code{\link{split.StdAssay}()},
\code{\link{subset.StdAssay}()}
}
\concept{stdassay}
\keyword{internal}
