% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{[[<-,Seurat,character,missing,StdAssay-method}
\alias{[[<-,Seurat,character,missing,StdAssay-method}
\title{Add Subobjects}
\usage{
\S4method{[[}{Seurat,character,missing,StdAssay}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \code{\link{Seurat}} object}

\item{i}{Name to add subobject as}

\item{j}{Ignored}

\item{...}{Ignored}

\item{value}{A valid subobject (eg. a \link[Assay]{v3} or
\link[Assay5]{v5} assay or a \link[DimReduc]{dimensional reduction})}
}
\value{
\code{x} with \code{value} added as \code{i}
}
\description{
Add Subobjects
}
\seealso{
See \link[=[[.Seurat]{here} for pulling subobjects using \code{[[},
\link[=$.Seurat]{here} for adding metadata with \code{[[<-}, and
\link[=[[<-,Seurat,NULL]{here} for removing subobjects and cell-level meta
data with \code{[[<-}

Seurat object, validity, and interaction methods
\code{\link{$.Seurat}()},
\code{\link{Seurat-class}},
\code{\link{Seurat-validity}},
\code{\link{[[.Seurat}()},
\code{\link{[[<-,Seurat,NULL}},
\code{\link{dim.Seurat}()},
\code{\link{dimnames.Seurat}()},
\code{\link{merge.Seurat}()},
\code{\link{names.Seurat}()},
\code{\link{subset.Seurat}()}
}
\keyword{internal}
