% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay5.R
\name{subset.Assay5}
\alias{subset.Assay5}
\title{Subset an Assay}
\usage{
\method{subset}{Assay5}(x, cells = NULL, features = NULL, layers = NULL, ...)
}
\arguments{
\item{x}{An \code{\link{Assay5}} object}

\item{cells}{Cell names}

\item{features}{Feature names}

\item{layers}{Layer to keep; defaults to all layers}

\item{...}{Ignored}
}
\value{
\code{x} with just the cells and features specified by
\code{cells} and \code{features} for the layers specified by \code{layers}
}
\description{
Subset an Assay
}
\seealso{
v5 Assay object, validity, and interaction methods:
\code{\link{$.Assay5}()},
\code{\link{Assay5-class}},
\code{\link{Assay5-validity}},
\code{\link{[.Assay5}()},
\code{\link{[[.Assay5}()},
\code{\link{dim.Assay5}()},
\code{\link{dimnames.Assay5}()},
\code{\link{merge.Assay5}()},
\code{\link{split.Assay5}()}
}
\concept{assay5}
