% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Top}
\alias{Top}
\title{Get the top}
\usage{
Top(data, num = 20, balanced = FALSE)
}
\arguments{
\item{data}{Data to pull the top from}

\item{num}{Pull top \code{num}}

\item{balanced}{Pull even amounts of from positive and negative values}
}
\value{
The top \code{num}
}
\description{
Get the top
}
\keyword{internal}
